### 2022-05-04 #4373 @editorialbot
[PRE REVIEW]: BifacialSimu: Hollistic Simulation of large-scale Bifacial Photovoltaic Systems
**Submitting author:** @EwaGomez (<a href="http://orcid.org/0000-0002-7826-3754">Eva-Maria Grommes</a>)
**Repository:** https://github.com/cire-thk/BifacialSimu
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
