### 2022-05-05 #4373 @kanderso-nrel
Howdy, I want to quickly point out a couple things:
- pvfactors and bifacial_radiance are both BSD-3 licensed, which I understand to mean their license texts must accompany any source code taken from their codebases.  Unless I missed it, neither of their license files are currently present in the bifacialsimu repo.
- Since pvlib is used in bifacialsimu itself as well as in the underlying bifacial simulation packages pvfactors and bifacial_radiance, it would be nice to cite it in the manuscript :) (see https://github.com/pvlib/pvlib-python#citing)
