### 2022-05-06 #4373 @kyleniemeyer
@EwaGomez while we are finding an editor to handle your submission, can you respond to / address the licensing concerns that @kanderso-nrel pointed out above? 👆
