### 2022-05-09 7b1607b
new trouble shooting and bifacial_radiance and viewfactor liscene integrated in the ReadMe
