### 2022-05-12 #4373 @sidihamady
@EwaGomez I would like to know if the authors use this code under Linux? The documentation is very Windows-centric, and it doesn't seem easy to use on Linux. Thanks.
