### 2022-05-12 #4373 @EwaGomez
@sidihamady We used the code under windows and Mac OS, but until now not under Linux. But I would like to consider that in the further development.
