### 2022-05-16 2e521a6
Adding Info Buttons Troubleshooting
Got an error running the code. Troubleshooting and error messages are shown in the comments in command lines 201 till 211. Python was basically unable to locate an image, which I was trying to integrate inside my info button.

an alternative info button was added just to double check the functionality of the written code with an integrate info image. (namely in line 213 and 214.

P.S. location of code lines are not optimal. will relocate them later. Having them at this level makes it easier to differentiate between frames in notebook.
