### 2022-05-16 7c58a99
Info Button Troubleshooting
The following changes removed the error message that was appearing previously; however the image did not appear on the Frame.

The background color of the frame was modified in line 189, to visualize if any changes were caused. as a result the button appear to be there as located in row 1 and column 2, however it is none functional and the image is not visible
