### 2022-05-18 #4373 @rkurchin
Sidi, don’t worry – this is still the pre-review issue; you’ll get a checklist for feedback on the review once we find another reviewer.

—————————————
Rachel C. Kurchin, PhD
Postdoctoral Researcher (formerly MolSSI and MFI Fellow)
Carnegie Mellon Mechanical Engineering and Materials Science
https://rkurchin.github.io
