+++ GUI.py
@@ -186,7 +186,7 @@ class Window(tk.Tk):
             
         #Creating the Frames
         #namecontrol_frame=tk.Frame(frame, width=200, height=60)
-        namecontrol_frame=tk.Frame(my_notebook, width=200, height=60,bg="blue") 
+        namecontrol_frame=tk.Frame(my_notebook, width=200, height=60) 
         #simulationMode_frame=tk.Frame(frame, width=200, height=60)
         simulationMode_frame=tk.Frame(my_notebook, width=200, height=60)
         simulationParameter_frame=tk.Frame(frame, padx=50, height=60)
@@ -194,27 +194,17 @@ class Window(tk.Tk):
         ModuleParameter_frame=tk.Frame(my_notebook, width=200, height=60)
         simulationFunction_frame=tk.Frame(frame, width=200, height=60)
         
+        # Defining the message box commmand function in Main Control tab
         def button1():
             messagebox.showinfo("Info!", "Text still to be added later")
         
-        
         #assigning the info Icon to a variable
-        self.infoButton= Image.open(rootPath+'\Lib\Button_Images\Button-Info-icon.png')
-        Info_image= self.infoButton
-        #inserting random info button to be modified later since below code line not functioning properly
-        Info1 = Button(namecontrol_frame, image=Info_image,command = button1, borderwidth=0).grid(row=1,column=2)
-
-        # =============================================================================
-        # Error Message from above commands in line 201 and 203
-        # 
-        # FileNotFoundError: [Errno 2] No such file or directory: 
-        # 'C:\\Users\\faris\\OneDrive\\Documents\\GitHub\\BifacialSimuLib\\Button logos\\Button-Info-icon.png'
-        # =============================================================================
+        namecontrol_frame.infoButton= PhotoImage(file= rootPath+'\Lib\Button_Images\Button-Info-icon.png')
+        Info_image= namecontrol_frame.infoButton
         
-        # inserting info button to Main Control frame without Image to double check position of button and functionality
-        # info_button_1 = Button(namecontrol_frame, text="INFO",command = button1, borderwidth=1).grid(row=1,column=2)
-        
-
+        #inserting info button in Main Control tab
+        Info1 = Button(namecontrol_frame, image=Info_image,command = button1, borderwidth=0)
+        Info1.grid(row=0,column=2)
 
         namecontrol_frame.bind("<Configure>", self._on_frame_configure)
         simulationMode_frame.bind("<Configure>", self._on_frame_configure)
