### 2022-05-20 #4373 @sidihamady
Dear @EwaGomez thanks for your quick reply.
I have some preliminary programming-related comments in this preview stage:
1. I think it would be important to offer a Linux version and I'm glad to know that you plan to do so. Technically it is quite feasible if we consider that the used libraries are OS-independent. Of course, the user can develop himself a Linux version, but if it's done, tested and maintained by the initial developers, it's better
2. Is it possible to include in the documentation a diagram or a description of the software architecture?
3. It looks like the simulations are started in the main thread (GUI), isn't it? This explains why the interface may freeze. Could you use the 'threading' library which is simple to use with Python? This will make it possible to launch long simulations in working threads (with the possibility of communicating with the GUI thread (via events for instance) e.g. to show the simulation progress) and to solve this issue.
4. There are still some untranslated comments and descriptions in German.
