### 2022-05-20 #4373 @EwaGomez
Dear @sidihamady  thank you for your comments!
Very helpful points! I will include a diagram of the software architecture in the ReadMe. I am also working on more description fields in the GUI. I am not exactly sure how the review works for JOSS, do I have to solve those issues before a possible publication?
I am continuously working on the software, but since I have a working version, which is already validaded, I wanted to publish this step, so other students/institutions/researchers, ... can use it.
