+++ Lib/Info_Messages/Simulation_Parameters.txt
@@ -1,0 +1,44 @@
+Single Axis Tracking:				adjusts the panels rotation around the horizantal axis to achieve higher efficiencies. 
+						[Once this is activated the "Limit Angle of Tracking System" should be set, which corresponds to the maximum angle that the modules can 
+						rotate to]
+
+Backtracking:					reduces the shadowing effects on modules during extreme low sun angles in the morning and evening. 
+						[a clearance height has to be entered which should represent distance between the module and the ground]
+
+Azimuth of PV surface:				orientation of solar panels [North=0°  East=90°  South=180°  West=270°]
+**Number of modules in row in X/Y-Axis
+**Number of sensors
+**Length of modules in X/Y-Axis
+
+Albedo Calculation: 				The calculation of the reflected diffused irradiation on the front side of the modules and almost the entire irradiance received by the rear side of 
+						the modules.
+
+Spectral Albedo Model: 				calculates the spectral albedo of the surface selected and is dependent not only on the spectral reflectance of the material (measured and
+                         					stored in a separate reflectivity file as a matrix of values) but also on the atmospheric irradiance and how they interact with each other.
+                         					This is an effective way of modeling the albedo for simulations; however it is quite computationally intensive and dependent on accurate
+                       					        weather data and comprehensive albedo values.
+						[Choosing the option Hourly Spectral Albedo requires a seperate Reflectivity file, which should be inserted under the Main Control tab]
+
+Simplified Variable Albedo Model:			The simplified variable albedo model uses the basic albedo value of a surface and implements a calculation based upon the position of the
+						sun and how this varies the albedo. This is a slightly more accurate solution to the constant albedo value however it is still heavily
+						dependent on the constant value.
+
+Hourly Albedo Measurement Data:			It is also possible to use previously measured hourly albedo values for further simulations, this can be useful for validation
+						bifacial gain of a new module within BifacialSimu or alternatively then a way to find out how other modules would function in this particular
+						scenario.
+						[the use of hourly measured albedo values is not possible with fixed tilt simulations at this point in time]
+
+Constant Albedo Value:				This is the simplest and therefore least accurate implementation of albedo in the simulation. It relies on a simple, average value of the
+						ground's albedo and inputs this into the irradiance calculation for the front and rear sides. These can be selected with the drop-down menu
+						according to what the surface is made of, or alternatively a custom numerical value can be entered.
+						[the name of the Surface will not be modified after entering a custom value]
+
+**Check Image in Main Control for reference
+
+
+
+
+
+
+
+
