### 2022-05-31 #4443 @editorialbot
[REVIEW]: BifacialSimu: Hollistic Simulation of large-scale Bifacial Photovoltaic Systems
**Submitting author:** @EwaGomez (<a href="http://orcid.org/0000-0002-7826-3754">Eva-Maria Grommes</a>)
**Repository:** https://github.com/cire-thk/BifacialSimu
**Branch with paper.md** (empty if default branch):
**Version:** 1.1.1
