### 2022-06-02 337c3ed
Added a figure with the simulation procedure to the Readme.md file
