### 2022-06-03 d4f6883
Messageboxes Updates
-adjusted messagebox size (height only) to fit in the display window

-added links to info boxes referring to Github Readme section
