### 2022-06-07 4f88375
changes in the contents of the new info box
