### 2022-06-11 9bcbf88
New integrated popup Info Messages
Since it is not possible to modify the messagebox functions provided by tkinter (like window size, button positions, numbers and names, fonts, backgroud etc.), I decided that it would be best to install our own message boxes using popup window functions.

Changes include:
-a new button named 'More Information' that opens the readme page of BifacialSimu on Github.
- popup window sized modified
- fixed the "Degree character: °" display, which was giving a display error on screen shown as "Â°"
