+++ GUI.py
@@ -32,6 +32,8 @@ import math
 
 import os
 import webbrowser
+from tkinter import *
+
 try:
     import tkinter as tk
     from tkinter import ttk
@@ -197,44 +199,129 @@ class Window(tk.Tk):
         ModuleParameter_frame=tk.Frame(my_notebook, width=200, height=60)
         simulationFunction_frame=tk.Frame(frame, width=200, height=60)
         
+        def github_infopage(option_1):
+            
+            if option_1 == "yes":
+                webbrowser.open("https://github.com/cire-thk/BifacialSimu#readme",new=1)
+            
+            
 # =============================================================================
 #         Defining the message box commmand functions:
 # =============================================================================
         # Main Control tab:
         def button_MC():
-            f_MC= open(rootPath+"\Lib\Info_Messages\Main_Control.txt")
-            text_MC= f_MC.read()
-            f_MC.close()
-            response=messagebox.askokcancel("Functions Info!", text_MC)
-            if response == 1:
-                webbrowser.open("https://github.com/cire-thk/BifacialSimu#readme",new=1)
-                
+            # f_MC= open(rootPath+"\Lib\Info_Messages\Main_Control.txt")
+            # text_MC= f_MC.read()
+            # f_MC.close()
+            # response=messagebox.askokcancel("Functions Info!", text_MC)
+            # if response == 1:
+            #     webbrowser.open("https://github.com/cire-thk/BifacialSimu#readme",new=1)
+            pop = Toplevel()
+            pop.title("Main Control Info!")
+            pop.geometry("1000x400")
+            pop.iconbitmap(rootPath+"\Lib\Button_Images\info_logo.ico")
+            
+            f= open(rootPath+"\Lib\Info_Messages\Main_Control.txt")
+            text_MP = f.read()
+            f.close()
+            # since ° is a unicode character it should be replaced by its character unicode to be read properly from .txt files
+            text_MP=text_MP.replace("Â°", "\u00b0")
+             
+            Info_label = Label(pop, text=text_MP,bg="white", font=("Arial",8),justify="left")
+            Info_label.pack(pady=5)
+            
+            info_frame= Frame(pop, bg="white")
+            info_frame.pack(pady=5)
+            
+            more_info_button = Button(info_frame,text='More Infomation', command= lambda:github_infopage("yes"))
+            more_info_button.grid(row=0,column=0)
+                                  
+            
+            
+                                  
         # Simulation Control tab
         def button_SC():
-            f_SC= open(rootPath+"\Lib\Info_Messages\Simulation_Control.txt")
-            text_SC= f_SC.read()
-            f_SC.close()
-            response=messagebox.askokcancel("Functions Info!", text_SC)
-            if response == 1:
-                webbrowser.open("https://github.com/cire-thk/BifacialSimu#readme",new=1)
-                
+            # f_SC= open(rootPath+"\Lib\Info_Messages\Simulation_Control.txt")
+            # text_SC= f_SC.read()
+            # f_SC.close()
+            # response=messagebox.askokcancel("Functions Info!", text_SC)
+            # if response == 1:
+            #     webbrowser.open("https://github.com/cire-thk/BifacialSimu#readme",new=1)
+               
+            pop = Toplevel()
+            pop.title("Simulation Control Info!")
+            pop.geometry("1400x700")
+            pop.iconbitmap(rootPath+"\Lib\Button_Images\info_logo.ico")
+            
+            f= open(rootPath+"\Lib\Info_Messages\Simulation_Control.txt")
+            text_MP = f.read()
+            f.close()
+            # since ° is a unicode character it should be replaced by its character unicode to be read properly from .txt files
[…]
