+++ Lib/Info_Messages/Simulation_Parameters.txt
@@ -1,13 +1,13 @@
-Single Axis Tracking:		adjusts the panels rotation around the horizantal axis to achieve higher efficiencies. 
-				[Once this is activated the "Limit Angle of Tracking System" should be set, which corresponds to the maximum angle that the modules can 
+Single Axis Tracking:		adjusts the panels rotation around the horizantal axis to achieve higher efficiencies.
+				[Once this is activated the "Limit Angle of Tracking System" should be set, which corresponds to the maximum angle that the modules can
 				rotate to]
 
-Backtracking:			reduces the shadowing effects on modules during extreme low sun angles in the morning and evening. 
-				[a clearance height has to be entered which should represent distance between the module and the ground]
+Backtracking:			reduces the shadowing effects on modules during extreme low sun angles in the morning and evening.
+				[a clearance height has to be entered which should represent distance between the module and the ground] 
 
-Azimuth of PV surface:		orientation of solar panels [North=0°  East=90°  South=180°  West=270°]
+Azimuth of PV surface:		orientation of solar panels [North=0°  East=90°  South=180°  West=270°]  
 
-Albedo Calculation: 		The calculation of the reflected diffused irradiation on the front side of the modules and almost the entire irradiance received by the rear side of 
+Albedo Calculation: 			The calculation of the reflected diffused irradiation on the front side of the modules and almost the entire irradiance received by the rear side of
 				the modules.
 
 Spectral Albedo Model: 		calculates the spectral albedo of the surface selected and is dependent not only on the spectral reflectance of the material (measured and
@@ -29,5 +29,3 @@ Constant Albedo Value:		This is the simplest and therefore least accurate implem
 				ground's albedo and inputs this into the irradiance calculation for the front and rear sides. These can be selected with the drop-down menu
 				according to what the surface is made of, or alternatively a custom numerical value can be entered.
 				[the name of the Surface will not be modified after entering a custom value]
-
-FOR MORE INFORMATION CHECK OUR README SECTION ON GITHUB OR CLICK THE "OK" BUTTON BELOW

[…]
