### 2022-06-14 4535041
Typo Fix in GUI for info box
