### 2022-06-16 cire-thk/BifacialSimu#7 @kanderso-nrel
[JOSS] Does the package have standard tests?
One of the JOSS review checklist items is about software testing:

>  Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

Having some kind of standard suite of tests has a lot of benefits for scientific software packages, including but not limited to:
- preventing unintentional changes to simulation results by checking that the software output doesn't change over time
- detecting incompatibilities with dependencies (e.g. perhaps tomorrow pvlib releases a new version that breaks your code)
- making it easier for new users to ensure they have installed the software correctly on their computer

What many other packages do, including pvlib, pvfactors, and bifacial_radiance, is have a suite of tests covering most or nearly all of the primary package code.  We also run these tests automatically for each commit to give us confidence that we are not accidentally breaking something.  Using pvlib as an example:
- test definitions: https://github.com/pvlib/pvlib-python/tree/master/pvlib/tests
- example automated test run: https://github.com/pvlib/pvlib-python/actions/runs/2503197724

In my opinion, testing is essentially a requirement for robust and reliable scientific software.  For pvlib we take it so seriously that nearly 100% of the package code is covered by tests (see https://app.codecov.io/gh/pvlib/pvlib-python) and we don't accept new code unless it has tests (see pvlib's [contributing docs](https://pvlib-python.readthedocs.io/en/stable/contributing.html#testing)).

Does BifacialSimu have anything along these lines?  I didn't see any tests, but maybe I missed it.  If not, do you have any plans to implement tests?

Cross-reference: https://github.com/openjournals/joss-reviews/issues/4443
