### 2022-06-16 cire-thk/BifacialSimu#8 @kanderso-nrel
[JOSS] Using modified versions of pvfactors and bifacial_radiance seems unsustainable
The installation instructions in the README say to install pvfactors and bifacial_radiance, then edit those installations to use modified files in this repository.  This approach seems problematic to me for a few reasons:

- It makes installation more manual and error-prone.
- It is brittle: the original projects will continue to evolve and diverge from the old code in this repository.  Maybe someday they will diverge enough that your modified code becomes incompatible with the latest version of pvfactors or bifacial_radiance and the code breaks.
- By replacing their code with yours, it means any improvements they make to those packages might not apply to your code.  For example maybe someday someone finds a bug in pvfactors's `pvground.py` and fixes it.  BifacialSimu won't get the bugfix because it is overwriting that file with an old version from before the bugfix.
- It isn't aligned with the open-source ethos of sharing improvements with the wider community.  Ideally, any improvements you make to a dependency would be contributed to the dependency itself so that everyone can benefit from it.

If you compare the bifacial_radiance `main.py` in this repository with the latest version in their repository ([here](https://github.com/NREL/bifacial_radiance/blob/development/bifacial_radiance/main.py)), there are already many differences.  I am worried that using modified files will not be sustainable in the long term, and maybe not even in the short term.

Would it be possible to either get rid of those changes, or to contribute those changes back to the original projects?  Maybe there is some other solution to this problem?

Cross-reference: https://github.com/openjournals/joss-reviews/issues/4443
