### 2022-06-16 cire-thk/BifacialSimu#9 @kanderso-nrel
[JOSS] No community guidelines
One of the JOSS review checklist items is about instructions for users and contributors:
> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I think these points are not really addressed in the README or anywhere else.  Here are some examples from other packages if you'd like some inspiration:
- pvlib: https://github.com/pvlib/pvlib-python#getting-support and https://pvlib-python.readthedocs.io/en/stable/contributing.html
- rdtools: https://rdtools.readthedocs.io/en/stable/developer_notes.html
- astropy: https://www.astropy.org/contribute.html

Cross-reference: https://github.com/openjournals/joss-reviews/issues/4443
