### 2022-06-16 cire-thk/BifacialSimu#10 @kanderso-nrel
[JOSS] paper.md feedback
Here are some notes I took while reading the `paper.md` manuscript file.  Hopefully this feedback is helpful!

## Statement of need

> PV is the third largest power generation technology and will play a significant role in future global electricity generation (International Renewable Energy Agengy, 2019).

I read this as claiming that solar is *currently* the third largest power generation technology, which I don't think is what the sentence was intended to convey.  Maybe something like "Current forecasts expect PV to be the third largest power generation technology by 2050..." would be better?

---

> There is a need for an open source software tool to simulate bifacial PV, which combines existing radiation and electrical models and provides a graphical user interface (GUI).

I think this deserves more explanation.  Why is the combination of viewfactor and raytracing desirable?  Why not just use one or the other?  I would also argue that albedo should be discussed here.  Why are advanced albedo models necessary?  I'm not saying they aren't necessary, but that it would strengthen the paper to discuss it.

---

This comment is a little more vague, but I as the reader am not sure what the main "selling point" of BifacialSimu is, or why I might want to use it instead of some other tool.  Here are the questions I'm asking myself as prospective user:

- Is it for more accurate system performance modeling? No, since it doesn't include all the auxiliary things needed to model system performance (soiling, inverters, etc).
- Is it to make bifacial irradiance modeling easier for non-programmers?  No, since other bifacial tools (bifacial_radiance, SAM) also have GUIs.
- Is it because it's open-source?  No, because the underlying irradiance models are also open-source.

Tied in with this is the question "why create a new tool instead of improving an existing tool?".  For example someone could say that it would have been better to add fancy albedo models to bifacial_radiance, and improve its existing gui, instead of making a new wrapper around it.  If you can, it would be good to clarify what exactly the need is and why these other existing tools don't address it.

## State of the field

> BifacialSimu offers the advantage of an integrated energy yield calculation, whereas bifacial_radiance need a third software, the open-source software PVMismatch [@Mikofski.2018].

I think this is no longer true; my understanding is that bifacial_radiance now includes its own PV I-V calculations using pvlib.  See for example their [Tutorial 21](https://github.com/NREL/bifacial_radiance/blob/main/docs/tutorials/21%20-%20Example%20Simulation%20-%20Modeling%20Module%20Performance%2C%20an%20End%20to%20End%20Simulation.ipynb).

---

>  The commonly known open-source software using the viewfactor model, pvfactors [@Anoma.2017] and bifacialVF [@Marion.2017], have both no energy yield calculation and no GUI, which limits the target audience and the application possibilities. The commonly used software PVsyst offers an energy yield calculation and a GUI, but the used physical models are only documented to a limited extent and the software must be paid [@PVsystSA].

NREL's [System Advisor Model](https://sam.nrel.gov/) has a viewfactor-based rear-side irradiance model, does energy yield calculation, has a GUI, is well-documented, and is open source, so if those are the relevant criteria then it's probably worth mentioning SAM here.

---

I also think BifacialSimu's advanced albedo modeling should be highlighted in this `State of the field` section, since that's a clear technical advantage compared to other packages.

---

Cross-reference: https://github.com/openjournals/joss-reviews/issues/4443
