### 2022-06-16 cire-thk/BifacialSimu#11 @kanderso-nrel
[JOSS] document all dependencies
One of the JOSS review checklist items is about dependencies:

> Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

I think right now the answer to that question is "no".  Looking through the package files, I see that, in addition to the dependencies explicitly mentioned in the README, BifacialSimu requires these external packages:

- pandas
- IPython
- matplotlib
- numpy
- tqdm
- webbrowser
- dateutil
- seaborn (optional)

I think at least some of these are not provided by the default Anaconda distribution, so they should be at least mentioned in the README installation instructions.

However, I agree with the checklist item that it would be better to use an automated package manager to track dependencies.  In the python world you can use `setup.py` or `setup.cfg` for this, which would make your package compatible with tools like `pip`.  You could also list it on pypi.org if you wanted.

Cross-reference: https://github.com/openjournals/joss-reviews/issues/4443
