### 2022-06-16 cire-thk/BifacialSimu#12 @kandersolar
[JOSS] git repository contains unnecessary files
There are a handful of `__pycache__` folders and `.DS_Store` files scattered around the repository.  Since these are unnecessary and essentially just clutter, I suggest removing them and creating a [.gitignore](https://git-scm.com/docs/gitignore) file to prevent them from coming back.

Cross-reference: https://github.com/openjournals/joss-reviews/issues/4443
