+++ GUI.py
@@ -165,7 +165,7 @@ ModuleDict = {
 class Window(tk.Tk):
     def __init__(self):
         tk.Tk.__init__(self)
-        self.geometry("1350x670")
+        self.geometry("1600x800")
         self.title('BifacialSimu')
         self.iconbitmap(rootPath+"\Lib\logos\App_icon.ico")
         yscroll = tk.Scrollbar(self, orient=tk.VERTICAL)
@@ -233,7 +233,7 @@ class Window(tk.Tk):
             info_frame= Frame(pop, bg="white")
             info_frame.pack(pady=5)
             
-            more_info_button = Button(info_frame,text='More Infomation', command= lambda:github_infopage("yes"))
+            more_info_button = Button(info_frame,text='More Information', command= lambda:github_infopage("yes"))
             more_info_button.grid(row=0,column=0)
                                   
             
@@ -265,7 +265,7 @@ class Window(tk.Tk):
             info_frame= Frame(pop, bg="white")
             info_frame.pack(pady=5)
             
-            more_info_button = Button(info_frame,text='More Infomation', command= lambda:github_infopage("yes"))
+            more_info_button = Button(info_frame,text='More Information', command= lambda:github_infopage("yes"))
             more_info_button.grid(row=0,column=0) 
             
         # Module Parameter tab
@@ -277,7 +277,7 @@ class Window(tk.Tk):
             # if response == 1:
             #     webbrowser.open("https://github.com/cire-thk/BifacialSimu#readme",new=1)
              pop= Toplevel()
-             pop.title("Main Control Info!")
+             pop.title("Module Parameter Info!")
              pop.geometry("1300x760")
              pop.iconbitmap(rootPath+"\Lib\Button_Images\info_logo.ico")
              f= open(rootPath+"\Lib\Info_Messages\Module_Parameter.txt")
@@ -292,7 +292,7 @@ class Window(tk.Tk):
              info_frame= Frame(pop, bg="white")
              info_frame.pack(pady=5)
              
-             more_info_button = Button(info_frame,text='More Infomation', command= lambda:github_infopage("yes"))
+             more_info_button = Button(info_frame,text='More Information', command= lambda:github_infopage("yes"))
              more_info_button.grid(row=0,column=0)
              
         # Simulation Parameter tab
@@ -320,7 +320,7 @@ class Window(tk.Tk):
             info_frame= Frame(pop_SP, bg="white")
             info_frame.pack(pady=5)
             
-            more_info_button = Button(info_frame,text='More Infomatino', command= lambda:github_infopage("yes"))
+            more_info_button = Button(info_frame,text='More Informatino', command= lambda:github_infopage("yes"))
             more_info_button.grid(row=0,column=0)
 # =============================================================================
 #         assigning the info Button Icon to a variable
@@ -1044,18 +1044,24 @@ class Window(tk.Tk):
             Entry_reflectivityfile.delete(0, END)
             Entry_reflectivityfile.insert(0, filename)   
             SimulationDict["spectralReflectancefile"]=Entry_reflectivityfile.get()
-        
+            
+        def Set_UTC_offset():
+            Longitude= float(Entry_longitude.get())
+            offset_result= round(Longitude*24/360)
+            Entry_utcoffset.delete(0,END)
+            Entry_utcoffset.insert(0, int(offset_result))
+                     
         #Changing the weatherfile
         Lab_weatherfile=ttk.Label(namecontrol_frame, text="Add Path of weatherfile:")
-        Lab_weatherfile.grid(row=4, column=0)
[…]
