### 2022-06-20 cire-thk/BifacialSimu#8 @EwaGomez
Thank you for your suggestion! I am already working on it. Unfortunatly the file changes are nesceary right now to apply our albedo models. I totally agree that it makes the installation more manual. My students and I will try to find a solution, it was already on our To-Do list, but I thought I would publish the software beforehand to make it open source and to get more suggestions. Thank you!
