### 2022-06-21 cire-thk/BifacialSimu#14 @sidihamady
[JOSS] ImportError: cannot import name '_missingKeyWarning' from 'bifacial_radiance.main'
When starting GUI.py in Spyder, I have:
**ImportError: cannot import name '_missingKeyWarning' from 'bifacial_radiance.main' (C:\Users\Sidi\anaconda3\lib\site-packages\bifacial_radiance\main.py)**
Removing it from Lib\site-packages\bifacial_radiance\module.py solves this issue but why? Is _missingKeyWarning really used?

[https://github.com/openjournals/joss-reviews/issues/4443#issuecomment-1158667882](https://github.com/openjournals/joss-reviews/issues/4443#issuecomment-1158667882)
