### 2022-06-21 cire-thk/BifacialSimu#15 @sidihamady
[JOSS] Simulations run in the GUI thread ==> freezing GUI
An issue that I mentioned in the Preview stage: the simulations run in the main GUI thread, which make it unresponsive during long time (even if the spyder console output is seen). Simulations should be run in a working thread using the 'threading' library, and a communication between the two threads can be done, e.g. to show the progress of calculations (like a progessbar in tkinter or a simple label). Using threading will make the GUI part remaining responsive and the simulations can be monitored and even stopped when needed. This makes the user experience much more easier.

[https://github.com/openjournals/joss-reviews/issues/4443#issuecomment-1158667882](https://github.com/openjournals/joss-reviews/issues/4443#issuecomment-1158667882)
