### 2022-06-21 cire-thk/BifacialSimu#16 @sidihamady
[JOSS] AUTHORS file
Please add an AUTHORS file to the repo, with the list you gave in Readme.md, with their contribution.
They can also be added in the acknowledgements section.

[https://github.com/openjournals/joss-reviews/issues/4443#issuecomment-1158667882](https://github.com/openjournals/joss-reviews/issues/4443#issuecomment-1158667882)
