### 2022-06-22 9917d78
Removed -/+ Sign from Latitude and Longitude Entries
Was causing a Floating point error while running simulations
