### 2022-06-23 cire-thk/BifacialSimu#12 @EwaGomez
Thank you for your valuable comment! We followed your suggestions and added a .gitignore file.
