### 2022-06-29 #4443 @EwaGomez
@rkurchin Thank you for checking in. I already solved one issue and I am working on the others! Am I supposed to solve all of them for the review or can I also explain why I think it's better to solve some later?
