### 2022-06-29 b5bf652
Create requirements.txt
lists all project dependencies which are required to be listed in the setup file
