### 2022-06-29 4295901
Create __init__.py
is a requirement for packaging the project.
Helps users that download this project to much easier call the functions/classes in this project, in other projects they are working on by using 'import BifacialSimu'
