+++ __init__.py
@@ -1,0 +1,1 @@
+from BifacialSimu.GUI import Window
