### 2022-06-29 786d70b
Create MANIFEST.in
to force the setup.py file include all folders/files that are desired to be packed in the project package.

could also be used to force the setup.py to exclude specific files/folders
