+++ MANIFEST.in
@@ -1,0 +1,10 @@
+graft Archiv 
+graft BifacialSimu 
+graft build 
+graft Data_conversion 
+graft Lib 
+graft Plots 
+graft ReflectivityData 
+graft Scripts 
+graft WeatherData 
+
