+++ bifacialsimu.egg-info/requires.txt
@@ -1,0 +1,13 @@
+pandas==1.3.4
+ipython==8.4.0
+matplotlib==3.4.3
+numpy==1.20.3
+Pillow==9.1.1
+pvlib==0.9.1
+python_dateutil==2.8.2
+pytz==2021.3
+requests==2.26.0
+seaborn==0.11.2
+Shapely==1.8.2
+tqdm==4.62.3
+bifacial_radiance==0.4.1
