+++ bifacialsimu.egg-info/SOURCES.txt
@@ -1,0 +1,152 @@
+LICENSE
+MANIFEST.in
+setup.py
+Archiv/030321_BiSim_mono_changed_Ghana.py
+Archiv/030321_BiSim_mono_changed_Ghana_variable - Kopie.py
+Archiv/090621_ZusammenführungFelix.py
+Archiv/090621_finalZusammenführung.py
+Archiv/180521_Zusammenführung.py
+Archiv/BiSim_main_NREL.py
+Archiv/BiSim_main_lemgo.py
+Archiv/BifacialSimu_spectralAlbedoHandler_Test.py
+Archiv/Dataframe_df_Cologne.csv
+Archiv/Dataframe_df_Golden.csv
+Archiv/GUI_Cologne.py
+Archiv/GUI_Ghana_Tamale.py
+Archiv/GUI_Golden.py
+Archiv/Measurement_Albedo.csv
+Archiv/mainBiSim.py
+Archiv/mainBiSim_NREL.py
+Archiv/mainBifacialSimu.py
+Archiv/module_alt.json
+Archiv/__pycache__/BifacialSimu_spectralAlbedoHandler_Test.cpython-39-pytest-7.0.0.pyc
+BifacialSimu/Controller/BifacialSimu_simulationController.py
+BifacialSimu/Controller/__pycache__/BifacialSimu_simulationController.cpython-39.pyc
+BifacialSimu/Handler/BifacialSimu_calculationHandler.py
+BifacialSimu/Handler/BifacialSimu_dataHandler.py
+BifacialSimu/Handler/BifacialSimu_radiationHandler.py
+BifacialSimu/Handler/BifacialSimu_spectralAlbedoHandler.py
+BifacialSimu/Handler/BifacialSimu_spectralAlbedoHandler_1_row.py
+BifacialSimu/Handler/__pycache__/BifacialSimu_calculationHandler.cpython-39.pyc
+BifacialSimu/Handler/__pycache__/BifacialSimu_dataHandler.cpython-39.pyc
+BifacialSimu/Handler/__pycache__/BifacialSimu_radiationHandler.cpython-39.pyc
+BifacialSimu/Handler/__pycache__/BifacialSimu_spectralAlbedoHandler.cpython-39.pyc
+BifacialSimu/Handler/__pycache__/BifacialSimu_spectralAlbedoHandler_1_row.cpython-39.pyc
+Data_conversion/10minute_Flughafen.csv
+Data_conversion/10minute_to_minute_radiation.py
+Data_conversion/15minute_to_hourly_albedo.py
+Data_conversion/Golden_Albedo_15Min.csv
+Data_conversion/Measurement_Albedo.csv
+Data_conversion/Measurement_GHI_DHI_DNI.csv
+Data_conversion/hourly_Albedo.csv
+Data_conversion/hourly_Albedo_Golden.csv
+Data_conversion/hourly_GHI_DHI_DNI.csv
+Data_conversion/minute_GHI_DHI.csv
+Data_conversion/minute_to_hourly_albedo.py
+Data_conversion/minute_to_hourly_radiation.py
+Data_conversion/minute_zenith.csv
+Data_conversion/minute_zenith.py
+Lib/Button_Images/Button-Info-icon.png
+Lib/Button_Images/info_logo.ico
+Lib/Info_Messages/Main_Control.txt
+Lib/Info_Messages/Module_Parameter.txt
+Lib/Info_Messages/Simulation_Control.txt
+Lib/Info_Messages/Simulation_Parameters.txt
+Lib/default/Example_Config.png
+Lib/default/Sarah_Cologne.ini
+Lib/default/Sarah_Golden.ini
+Lib/default/default.ini
[…]
