### 2022-06-29 cire-thk/BifacialSimu#16 @EwaGomez
Thank you for your suggestion! I already mentioned the students in the Acknowledgement section "A special thanks to all the students of the University of Applied Sciences Cologne who have contributed so far to this project, it would not have been as successful without their input." - BifacialSimu is part of my Dissertation and research and there are many students projects helping to develop it further - all are supervised by me very closly. Do you still think I should add an authors file additionally to the Readme?
