### 2022-06-29 2b4c2de
added Getting Support section to ReadMe
