### 2022-06-29 d4ed4c5
small edits to setup.py
