+++ setup.py
@@ -20,7 +20,7 @@ setup(
     
 
     # Define the library name, this is what is used along with `pip install`.
-    name='bifacialsimu',
+    name='BifacialSimu',
 
     # Define the author of the repository.
     author='Eva-Maria Grommes',
@@ -37,7 +37,7 @@ setup(
 
     # Here is a small description of the library. This appears
     # when someone searches for the library on https://pypi.org/search.
-    description='Simualtion tool for energies released by bifacial photovoltaic models',
+    description='Hollistic Simulation of large-scale Bifacial Photovoltaic Systems',
 
     # I have a long description but that will just be my README
     # file, note the variable up above where I read the file.
@@ -63,11 +63,12 @@ setup(
         'seaborn==0.11.2',
         'Shapely==1.8.2',
         'tqdm==4.62.3',
-        'bifacial_radiance==0.4.1',#content changed!      
+        'bifacial_radiance==0.4.1',#content changed!    
+        #und pvfactors - welche Version?  
     ],
 
     # Here are the keywords of my library.
-    keywords='Bifacialsimu, bifacial simulation,  energy yield of bifacial PV technology',
+    keywords='Bifacialsimu, bifacial simulation,  energy yield of bifacial PV systems, albedo',
 
     # here are the packages I want "build."
     packages=find_packages(
