### 2022-06-30 cire-thk/BifacialSimu#16 @sidihamady
Dear @EwaGomez, in my opinion, a participant is not an author if his contribution was only to test/debug the software, without having had a contribution in the design and actual writing of the code. In this case, he/she is thanked in the Acknowledgements section. Otherwise, he is considered co-author of the software. My colleague @kanderso-nrel can give us his opinion. In any case, you are the major author of the code and after this clarification I will validate the point "Contribution and authorship".
