### 2022-06-30 cire-thk/BifacialSimu#16 @kanderso-nrel
Perhaps the distinction between authors of the paper and authors of the code is relevant here.  JOSS provides guidance on the former:
- https://joss.readthedocs.io/en/latest/submitting.html#authorship
- https://joss.readthedocs.io/en/latest/review_criteria.html#authorship

I don't have any objection to the authors list currently in the paper.  As for a separate AUTHORS file in the repository: I think a list that must be updated manually is more work and more likely to be neglected than GitHub's built-in contributor tracking feature is.  However it is nice in some cases to be able to list specific contributions, or in case you want to define a "contributor" more broadly than GitHub does.  Regardless I don't feel strongly about whether such a file should be created in this case.
