### 2022-07-05 #4443 @rkurchin
Hi again, just checking in with author @EwaGomez regarding the issues already filed and reviewers @sidihamady and @kanderso-nrel regarding progress of their reviews!
