### 2022-07-07 c9e2d95
Setup.py
changes include:
- packing the BifacialSimu folders using setup.py
- moving all source files and folder into 'BifacialSimu_src' folder, and keeping everything relating to packaging and distribution outside the source folder (better organizing purposes)
