+++ build/lib/BifacialSimu_src/Scripts/pvfactors/V_1_5_1/pvarray.py
@@ -1,0 +1,272 @@
+"""Module containing PV array classes, which will use PV rows and ground
+geometries."""
+
+import numpy as np
+from pvfactors.config import X_ORIGIN_PVROWS, DISTANCE_TOLERANCE
+from pvfactors.geometry.base import \
+    _get_solar_2d_vectors, BasePVArray, _get_rotation_from_tilt_azimuth
+from pvfactors.geometry.pvrow import TsPVRow
+from pvfactors.geometry.pvground import TsGround
+
+
+class OrderedPVArray(BasePVArray):
+    """An ordered PV array has a flat horizontal ground, and pv rows which
+    are all at the same height, with the same surface tilt and azimuth angles,
+    and also all equally spaced. These simplifications allow faster and easier
+    calculations. In the ordered PV array, the list of PV rows must be
+    ordered from left to right (along the x-axis) in the 2D geometry."""
+
+    y_ground = 0.  # ground will be at height = 0 by default
+
+    def __init__(self, axis_azimuth=None, gcr=None, pvrow_height=None,
+                 n_pvrows=None, pvrow_width=None, param_names=None,
+                 cut=None, x_min=None, x_max=None):                             #THKoeln
+        """Initialize ordered PV array.
+        List of PV rows will be ordered from left to right.
+
+        Parameters
+        ----------
+        axis_azimuth : float, optional
+            Azimuth angle of rotation axis [deg] (Default = None)
+        gcr : float, optional
+            Ground coverage ratio (Default = None)
+        pvrow_height : float, optional
+            Unique height of all PV rows in [m] (Default = None)
+        n_pvrows : int, optional
+            Number of PV rows in the PV array (Default = None)
+        pvrow_width : float, optional
+            Width of the PV rows in the 2D plane in [m] (Default = None)
+        param_names : list of str, optional
+            List of surface parameter names for the PV surfaces
+            (Default = None)
+        cut : dict, optional
+            Nested dictionary that tells if some PV row sides need to be
+            discretized, and how (Default = None).
+            Example: {1: {'front': 5}}, will create 5 segments on the front
+            side of the PV row with index 1
+        """
+        # Initialize base parameters: common to all sorts of PV arrays
+        super(OrderedPVArray, self).__init__(axis_azimuth=axis_azimuth)
+
+        # These are the invariant parameters of the PV array
+        self.gcr = gcr
+        self.height = pvrow_height
+        self.distance = (pvrow_width / gcr
+                         if (pvrow_width is not None) and (gcr is not None)
+                         else None)
+        self.width = pvrow_width
+        self.n_pvrows = n_pvrows
+        self.param_names = [] if param_names is None else param_names
+        self.cut = {} if cut is None else cut
+        self.x_min = x_min                                           #THKoeln
+        self.x_max = x_max                                           #THKoeln
+
+        # These attributes will be updated at fitting time
+        self.solar_2d_vectors = None
+        self.n_states = None
+        self.has_direct_shading = None
+        self.rotation_vec = None
+        self.shaded_length_front = None
+        self.shaded_length_back = None
+
+        # These attributes will be updated by the engine
+        self.ts_vf_matrix = None
+        self.ts_vf_aoi_matrix = None
+
+    @classmethod
+    def init_from_dict(cls, pvarray_params, param_names=None):
+        """Instantiate ordered PV array from dictionary of parameters
+
+        Parameters
+        ----------
+        pvarray_params : dict
+            The parameters defining the PV array
+        param_names : list of str, optional
+            List of parameter names to pass to surfaces (Default = None)
+
+        Returns
+        -------
[…]
