+++ BifacialSimu_src/BifacialSimu/Handler/BifacialSimu_spectralAlbedoHandler_1_row.py
@@ -1,0 +1,461 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Wed Sep 15 17:53:38 2021
+
+@author:        
+    Eva-Maria Grommes
+    Sarah Glaubitz
+
+Additional co-authors can be found here:
+https://github.com/cire-thk/bifacialSimu    
+
+name: BifacialSimu - spectralAlbedoHandler
+
+overview:
+    Calculation of spectral Albedo regarding the paper: 
+    "A Comprehensive Albedo Model for Solar Energy Applications: Geometric Spectral Albedo"
+    from Hesan Ziar, Furkan Fatih Sönmez, Olindo Isabella, and Miro Zeman; October 2019
+    
+"""
+
+import pandas as pd
+import numpy as np
+import math
+import dateutil.tz
+import datetime
+import csv
+import matplotlib.pyplot as plt
+from pvlib import spectrum, irradiance, atmosphere
+from pvfactors.geometry import OrderedPVArray
+from pvfactors.viewfactors.vfmethods import VFTsMethods
+from pvfactors.config import DISTANCE_TOLERANCE
+
+
+def getReflectivityData(simulationDict):
+    '''
+    Read the spectral reflectance data of the material (sand); R(lamda)
+    
+    Parameters
+    ----------
+    simulationDict: simulation Dictionary, which can be found in GUI.py
+    
+    Returns
+    -------
+    R_lamda: array of reflectvity values
+    '''
+    
+    # numpy array with reflectivity values, only colume 2 of the csv is read
+    R_lamda = np.genfromtxt(simulationDict['spectralReflectancefile'], delimiter=';', skip_header = 1, usecols=(1))
+   
+    return R_lamda
+    
+def modellingSpectralIrradiance(simulationDict, dataFrame, j):
+    '''
+    Model the spectral distribution of irradiance based on atmospheric conditions. 
+    The spectral distribution of irradiance is the power content at each wavelength 
+    band in the solar spectrum and is affected by various scattering and absorption 
+    mechanisms in the atmosphere.
+    
+    based on SPECTRL2 NREL Technical Report:
+    Bird, R, and Riordan, C., 1984, "Simple solar spectral model for direct and 
+    diffuse irradiance on horizontal and tilted planes at the earth's surface 
+    for cloudless atmospheres", NREL Technical Report TR-215-2436 
+    doi:10.2172/5986936.
+    
+    Parameters
+    ----------
+    simulationDict: simulation Dictionary, which can be found in GUI.py
+    dataFrame: pandas dataframe, which contains the weather data
+    j: current loop number, which represent the hour after starthour = index of df
+    
+    Returns
+    -------
+    spectra: dict of arrays with wavelength; dni_extra; dhi; dni; poa_sky_diffuse; poa_ground_diffuse; poa_direct; poa_global
+    '''
+       
+    df = dataFrame
+     
+    tilt = 0                                # [deg] always 0, because the ground is never tilted
+    azimuth = simulationDict['azimuth']     # [deg] same azimuth for ground surface as for PV panel
+    pressure = (df.iloc[j]['pressure']*100) # [Pa] air pressure; df value is in mbar so multiplied by 100 to Pa
+    water_vapor_content = 1.551             # [cm] Atmospheric water vapor content; data from AERONET for FZ Juelich for Sep 2021; Level 2 Quality
+    tau500 = 0.221                          # [-] aerosol optical depth at wavelength 500 nm; data from AERONET for FZ Juelich for Sep 2021; Level 2 Quality
+    ozone = 0.314                           # [atm-cm] Atmospheric ozone content; data from WOUDC for Aug 2021 for Hohenpeissenberg
[…]

[…]
