### 2022-07-11 cire-thk/BifacialSimu#11 @EwaGomez
Thank you for your comment!
I created a requirement.txt to display all dependencies.

Additionally we created a setup.py, which we are still running some tests with. That will make the installation process a lot easier. I am sure, we will be able to commit it in the next 14 days.
