### 2022-07-11 cire-thk/BifacialSimu#8 @EwaGomez
Dear @kanderso-nrel and @sidihamady ,

Unfortunately, the changes in pyground.py and main.py are necessary. My idea now is to implement copies of pvfactors and bifacial_radiance into BifacialSimu to integrate them into the pip-installation process. In those copies, the necessary changes are already done, so users do not have to exchange files themselves, which would prevent errors.

Doing so, when there is an update in bifacial_radiance or pvfactors, I could transfer my changes to the newer versions and test them. In the long run, I would like to contribute my changes to bifacial_radiance and pvfactors, but for now, I think that would be a good solution.

Is that suitable for you?
