### 2022-07-11 cire-thk/BifacialSimu#11 @kanderso-nrel
Thanks @EwaGomez, I suggest also mentioning in the README how to use the new requirements file during installation, unless you'd rather wait until the `setup.py` is ready.
