### 2022-07-11 cire-thk/BifacialSimu#8 @kanderso-nrel
I think that would address the first and second bullets in my list above.  Perhaps it would be helpful to open an issue in each repository describing your changes?  Maybe there is a better solution that the maintainers of those packages could suggest, and they could also indicate whether they would be interested in a PR for those changes, even if it will happen someday instead of today.

One other thing is I think you would have to also start tracking all of the dependencies of pvfactors and bifacial_radiance in addition to your own dependencies (#11).  FYI this practice of embedding one project within another is sometimes called "vendoring", in case you want to search for advice on this topic.
