### 2022-07-11 cire-thk/BifacialSimu#11 @EwaGomez
Yes, I will mention that when the setup.py is ready!  Thank you!
