### 2022-07-15 1146699
Removing old distribution folders and updating .gitignore
