+++ build/lib/BifacialSimu_src/Archiv/GUI_Golden.py
@@ -1,1790 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Mon Jun  7 11:39:16 2021
-@author:        
-    CIRE TH Cologne
-    Eva-Maria Grommes
-    Jan Schmitt
-    Sarah Glaubitz
-name:
-    Bifacial Simu - GUI
-overview:
-    Import of needed modules and paths.
-    Input of variables and settings for the bifacial simulation of PV-Modules 
-    with View Factors and/or Ray Tracing method. 
-    Command to run the simulation
-    
-last changes:
-    07.06.21 created
-"""
-    
-# Start-Befehl für Simulation
-
-# Später: Aufruf der GUI
-
-# -*- coding: utf-8 -*-
-
-# Import modules
-#import matplotlib
-#matplotlib.use("TkAgg")
-import sys
-import math
-
-import os
-try:
-    import tkinter as tk
-    from tkinter import ttk
-    from tkinter import filedialog
-    from tkinter import messagebox
-except:
-    import Tkinter as tk
-    import ttk
-
-from configparser import ConfigParser
-#import bifacial_radiance
-#import warnings
-
-
-from matplotlib import pyplot as plt
-import matplotlib.dates as dates
-import json
-from PIL import ImageTk, Image
-import datetime
-import csv
-import numpy as np
-import pandas as pd
-#import time
-#import pickle
-
-
-# aliases for Tkinter functions
-END = tk.END
-W = tk.W
-Entry = tk.Entry
-Button = tk.Button
-Radiobutton = tk.Radiobutton
-IntVar = tk.IntVar
-PhotoImage = tk.PhotoImage
-
-# Importieren der nötigen Module und Pfade
-# Path handling
-rootPath = os.path.realpath(".")
-print(rootPath)
-# Include paths
-sys.path.append(rootPath + "/BifacialSimu/Controller")
-sys.path.append(rootPath + "/BifacialSimu/Handler")
-
-# Include modules
-import BifacialSimu_simulationController
-import BifacialSimu_dataHandler
-
-
-
-# Eingabe von Simulationsvariablen und Einstellungen für die Simulationsmethode
-
-"""
-        Sets the mode for simulation: str
-        mode 1 : front simulation with PVfactors, back simulation with Raytracing
-        mode 2 : front and back simulation with Viewfactors
-        mode 3 : front and back simulation with Raytracing
-        mode 4 : only front simulation with Viewfactors
-        mode 5 : only back simulation with Raytracing
-"""
-
-# simulation parameters and variables
-SimulationDict = {
-'simulationName' : 'NREL_best_field_row_2',
-'simulationMode' : 2, 
-'localFile' : True, # Decide wether you want to use a  weather file or try to download one for the coordinates
-'weatherFile' : (rootPath +'/WeatherData/Golden_USA/SRRL_Golden_Jul_2020.csv'), # weather file in TMY format 
-'spectralReflectancefile' : (rootPath + '/ReflectivityData/grass_interpolated.csv'),
-'cumulativeSky' : False, # Mode for RayTracing: CumulativeSky or hourly
-'startHour' : (2020, 7, 6, 0),  # Only for hourly simulation, yy, mm, dd, hh
-'endHour' : (2020, 7, 22, 0),  # Only for hourly simulation, yy, mm, dd, hh
-'utcOffset': -7,
-'tilt' : 30, #tilt of the PV surface [deg]
-'singleAxisTracking' : True, # singleAxisTracking or not
-'backTracking' : False, # Solar backtracking is a tracking control program that aims to minimize PV panel-on-panel shading 
[…]
