### 2022-07-15 e03ccb3
creating a Vendor folder with Bifacial_radiance and pvfactors modified packages + changing project dependency paths
With this folder built in, users would stop to manually relocate the modified files from bifacial_radiance and pvfactors packages.
