### 2022-07-15 819579e
new packaging including Vendoring
Packaging folders are still essential to upload to github in order for the setup.py to function properly with the 'pip install .' command.

Only when BifacialSimu is published on PyPi these packages could be removed from github, und users will then be able to use the 'pip install Bifacialsimu' instead
