### 2022-07-15 e4cd0da
removing packaging from Github and updating .gitignore again
