+++ build/lib/BifacialSimu_src/Scripts/pvfactors/V_1_5_1/pvground.py
@@ -1,1095 +1,0 @@
-"""Classes for implementation of ground geometry"""
-from pvfactors import PVFactorsError
-from pvfactors.config import (
-    MAX_X_GROUND, MIN_X_GROUND, Y_GROUND, DISTANCE_TOLERANCE, COLOR_DIC)
-from pvfactors.geometry.base import (
-    BaseSide, PVSegment, ShadeCollection, PVSurface)
-from pvfactors.geometry.timeseries import (
-    TsShadeCollection, TsLineCoords, TsPointCoords, TsSurface,
-    _get_params_at_idx)
-from shapely.geometry import LineString
-import numpy as np
-from copy import deepcopy
-
-
-class TsGround(object):
-    """Timeseries ground class: this class is a vectorized version of the
-    PV ground geometry class, and it will store timeseries shaded ground
-    and illuminated ground elements, as well as pv row cut points."""
-
-    # TODO: this needs to be passed at initialization for flexibility
-    #x_min = MIN_X_GROUND        #THKoeln
-    #x_max = MAX_X_GROUND        #THKoeln
-
-    def __init__(self, shadow_elements, illum_elements, param_names=None,
-                 flag_overlap=None, cut_point_coords=None, y_ground=None):
-        """Initialize timeseries ground using list of timeseries surfaces
-        for the ground shadows
-
-        Parameters
-        ----------
-        shadow_elements : \
-        list of :py:class:`~pvfactors.geometry.pvground.TsGroundElement`
-            Timeseries shaded ground elements
-        illum_elements : \
-        list of :py:class:`~pvfactors.geometry.pvground.TsGroundElement`
-            Timeseries illuminated ground elements
-        param_names : list of str, optional
-            List of names of surface parameters to use when creating geometries
-            (Default = None)
-        flag_overlap : list of bool, optional
-            Flags indicating if the ground shadows are overlapping, for all
-            time steps (Default=None). I.e. is there direct shading on pv rows?
-        cut_point_coords : \
-        list of :py:class:`~pvfactors.geometry.timeseries.TsPointCoords`, \
-        optional
-            List of cut point coordinates, as calculated for timeseries PV rows
-            (Default = None)
-        y_ground : float, optional
-            Y coordinate of flat ground [m] (Default=None)
-        """
-        # Lists of timeseries ground elements
-        self.shadow_elements = shadow_elements
-        self.illum_elements = illum_elements
-        # Shade collections
-        list_shaded_surf = []
-        list_illum_surf = []
-        for shadow_el in shadow_elements:
-            list_shaded_surf += shadow_el.all_ts_surfaces
-        for illum_el in illum_elements:
-            list_illum_surf += illum_el.all_ts_surfaces
-        self.illum = TsShadeCollection(list_illum_surf, False)
-        self.shaded = TsShadeCollection(list_shaded_surf, True)
-        # Other ground attributes
-        self.param_names = [] if param_names is None else param_names
-        self.flag_overlap = flag_overlap
-        self.cut_point_coords = [] if cut_point_coords is None \
-            else cut_point_coords
-        self.y_ground = y_ground
-        self.shaded_params = dict.fromkeys(self.param_names)
-        self.illum_params = dict.fromkeys(self.param_names)
-
-    @classmethod
-    def from_ts_pvrows_and_angles(cls, list_ts_pvrows, alpha_vec, rotation_vec,
-                                  y_ground=Y_GROUND, flag_overlap=None,
-                                  param_names=None,x_min=MIN_X_GROUND, x_max=MAX_X_GROUND):      #THKoeln
-        """Create timeseries ground from list of timeseries PV rows, and
-        PV array and solar angles.
-
[…]
