+++ build/lib/BifacialSimu_src/BifacialSimu/Handler/BifacialSimu_dataHandler.py
@@ -1,155 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Mon Jun  7 11:39:16 2021
-@author:        
-    Eva-Maria Grommes
-
-Additional co-authors can be found here:
-https://github.com/cire-thk/bifacialSimu    
-
-name:
-    BifacialSimu - dataHandler
-
-overview:
-    Handles the import and export of data 
-    used by the bifacial simulation of PV-Modules
-
-"""
-
-
-# Klasse zum Importieren/Umwandeln von benötigten Daten
-    
-    # Funktion zum einlesen/umwandeln von Exceldateien?    
-
-
-# Klasse zum Exportieren/Umwandeln von benötigten Daten
-
-    # Funktion zum zusammenführen von Radiaton-Daten
-    # Funktion zum Exportieren von Radiation-Daten
-    # Funktion zum Exportieren von Grafiken
-    # Funktion zum Exportieren eines Berichts?
- 
-
-
-import datetime
-import pandas as pd
-import os #to import directories
-import sys
-import dateutil
-import numpy as np
-
-
-
-
-
-
-# Path handling
-rootPath = os.path.dirname(os.path.dirname(os.path.realpath(".")))
-
-import BifacialSimu_radiationHandler
-
-class DataHandler:
-    def __init__(self):
-        self.localDir = rootPath
-
-    
-    def setDirectories(self, outputFolder = "Outputs"):
-        '''
-        outputFolder: str
-            Name of output folder (will be created)
-        
-        '''
-  
-        # Path to import irradiance data / Quality for plot export
-        now = datetime.datetime.now()
-        date_time = now.strftime("%Y %m %d_%H_%M") # get current date and time
-        outputPath = os.path.join(self.localDir, outputFolder)
-        resultsPath = os.path.join(outputPath, date_time + '_results/' ) 
-        if not os.path.exists(resultsPath):
-            os.makedirs(resultsPath)        # create path to output
-        
-        return resultsPath
-    
-    
-    def getWeatherData(self, simulationDict, resultsPath):
-        """
-        Function to create a Radiance Obj with bifacial_radiance and read weather data.
-        Can read EPW weather files from input location data or local weather files
-        
-        Parameters
-        ----------
-        simulationDict: simulation Dictionary, which can be found in BifacialSimuu_main.py
-        resultsPath: output filepath
-        """
-        
-        demo = BifacialSimu_radiationHandler.RayTrace.createDemo(simulationDict, resultsPath)
-        if simulationDict['localFile'] == False:
-            try:
-                longitude = simulationDict['longitude']
-                latitude = simulationDict['latitude']
-            
-                epwfile = demo.getEPW(latitude,longitude) # pull TMY data for any global lat/lon
-            except ConnectionError: # no connection to automatically pull data
-                pass
-            metdata = demo.readEPW(epwfile) # read in the EPW weather data from above
-        else:
-            metdata = demo.readTMY(simulationDict['weatherFile'])
-        
-        return metdata, demo
-    
-    def passEPWtoDF(self, metdata, simulationDict, resultsPath):
-        
-        """
-        Function to pass irradiance data and temperature from metdata Object created by bifacial_radiance to a pandas dataframe.
-        The dataframe will be further used in the class ViewFactors.
-        Additionally, a timeindex will be set.
-        
-        Parameters
-        ----------
-        simulationDict: simulation Dictionary, which can be found in BifacialSimu_main.py
-        metdata: Object containing meteorological data and sun parameters        
-        resultsPath: output filepath       
-        """
-        df = metdata.solpos
-
-        df['ghi'] = metdata.ghi
[…]
