+++ BifacialSimu_src/Archiv/030321_BiSim_mono_changed_Ghana_variable - Kopie.py
@@ -1,717 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-Created on Mon Oct  7 12:57:54 2019
-
-@author:        Eva-Maria Grommes
-Bifacial photovoltaic yield simulation as a function of the albedo
-
-
-Last change:
-    23.02.21 New solarposition calc method, new energy density calc
-    09.03.21 Variable Albedo implemented
-
-"""
-
-# Import
-from IPython import get_ipython
-get_ipython().magic('reset -sf')
-
-import pandas as pd #pandas = can read .csv as input
-import matplotlib.pyplot as plt #display shadows
-import numpy as np
-import os #to import directories
-import warnings
-import math
-import pvlib #for electrical output simulation
-from datetime import datetime #get current date and time
-from pvfactors.viewfactors.aoimethods import faoi_fn_from_pvlib_sandia #to calculate AOI reflection losses
-from pvfactors.engine import PVEngine
-from pvfactors.irradiance import HybridPerezOrdered
-from pvfactors.geometry import OrderedPVArray
-from pvfactors.viewfactors import VFCalculator
-from tqdm import tqdm
-
-# Settings for calculating ViewFactor
-get_ipython().run_line_magic('matplotlib', 'inline')
-np.set_printoptions(precision=3, linewidth=300)
-warnings.filterwarnings('ignore')
-
-# seaborn makes your plots look better
-try:
-    import seaborn as sns
-    sns.set(rc={"figure.figsize": (12, 6)})
-    sns.set_style("darkgrid") # Other available styles: whitegrid, dark, white, ticks
-    sns.set_context("paper") # Scaling the axes, also available: talk, poster, 
-    sns.set_palette("muted") 
-    sns.set_color_codes()
-except ImportError:
-    print('We suggest you install seaborn using conda or pip and rerun this cell')
-
-#########################################################
-
-# Path to import irradiance data / Quality for plot export
-LOCAL_DIR = os.getcwd()
-DATA_DIR = os.path.join(LOCAL_DIR, 'C:/Users/egrommes/sciebo/Simulation/Ghana_mono_changed - variable Albedo')
-filepath = os.path.join(DATA_DIR, 'Ghana_Reference.csv')
-dpi = 150 #Quality for plot export
-
-# Pfad Sarah: C:/Users/sarah/OneDrive/Desktop/Hiwi/Simulation/Ghana_mono_changed
-# Pfad Eva: C:/Users/egrommes/sciebo/Simulation/Ghana_mono_changed
-#########################################################
-    
-# Dictionary for Input Parameters
-simulationParameter = {
-    'n_pvrows': 3, #number of PV rows
-    'number_of_segments': 5, #number of segments for each PVrow
-    'pvrow_height': 2.95, #height of the PV rows, measured at their center [m]
-    'pvrow_width': 1.968, #width of the PV panel in row, considered 2D plane [m]
-    'pvmodule_width': 0.992, #length of the PV panel in row, considered 2D plane [m]
-    'surface_azimuth': 135, #azimuth of the PV surface [deg] 90°= East, 135° = South-East, 180°=South
-    'surface_tilt': 35, #tilt of the PV surface [deg]
-    #'albedo': 0.259597496, # Measured Albedo average value
-    'a0': 0.259597496, # Measured Albedo under direct illumination with a solar zenith angle of approx. 60°
-    'adiff': 0.22, # Measured Albedo under 100% diffuse illumination
[…]

[…]
