### 2022-07-18 ae9852a
Changes in ReadMe with new installation process
