### 2022-07-18 cire-thk/BifacialSimu#8 @EwaGomez
Thank you for your advice with the vendoring! We did as you said and made the installation process a lot easier using a setup.py and pip install . The changes are also integrated into the ReadMe file as well as the requirements. Doing so, we changed the vendored version of bifacial_radiance to 0.4.1.

I will definitely open an issue in each repository with my changes and ask them for advice. Thank you!
