### 2022-07-18 cire-thk/BifacialSimu#14 @EwaGomez
Thank you for your comment! You are totally right, we don't need _missingKeyWarning anymore and removed it out of the current version.
