### 2022-07-18 cire-thk/BifacialSimu#15 @EwaGomez
Thank you for your comment! I will read into the threading library!
Would you mind if I resolve this issue after the JOSS review? Right now I would love to concentrate on more research related issues and later in the year on a more user-friendly design.
