### 2022-07-21 eba99da
Installation changes in GUI, translation changes in calculation and handlers
small changes in installation procedure; comment translation from german bits to english
