### 2022-07-25 #4443 @EwaGomez
Hi @rkurchin, I was already able to close 7 issues and I am working on the last 3. The discussion with both of the reviewers is really helpful and I have never experienced such a good open review process. Thanks a lot!
