+++ BifacialSimu.egg-info/SOURCES.txt
@@ -9,30 +9,18 @@ BifacialSimu.egg-info/top_level.txt
 BifacialSimu_src/BifacialSimu_main.py
 BifacialSimu_src/GUI.py
 BifacialSimu_src/__init__.py
-BifacialSimu_src/Archiv/030321_BiSim_mono_changed_Ghana.py
-BifacialSimu_src/Archiv/030321_BiSim_mono_changed_Ghana_variable - Kopie.py
-BifacialSimu_src/Archiv/090621_ZusammenführungFelix.py
-BifacialSimu_src/Archiv/090621_finalZusammenführung.py
-BifacialSimu_src/Archiv/180521_Zusammenführung.py
-BifacialSimu_src/Archiv/BiSim_main_NREL.py
-BifacialSimu_src/Archiv/BiSim_main_lemgo.py
-BifacialSimu_src/Archiv/BifacialSimu_spectralAlbedoHandler_Test.py
-BifacialSimu_src/Archiv/Dataframe_df_Cologne.csv
-BifacialSimu_src/Archiv/Dataframe_df_Golden.csv
-BifacialSimu_src/Archiv/GUI_Cologne.py
-BifacialSimu_src/Archiv/GUI_Ghana_Tamale.py
-BifacialSimu_src/Archiv/GUI_Golden.py
-BifacialSimu_src/Archiv/Measurement_Albedo.csv
-BifacialSimu_src/Archiv/mainBiSim.py
-BifacialSimu_src/Archiv/mainBiSim_NREL.py
-BifacialSimu_src/Archiv/mainBifacialSimu.py
-BifacialSimu_src/Archiv/module_alt.json
 BifacialSimu_src/BifacialSimu/Controller/BifacialSimu_simulationController.py
+BifacialSimu_src/BifacialSimu/Controller/__pycache__/BifacialSimu_simulationController.cpython-39.pyc
 BifacialSimu_src/BifacialSimu/Handler/BifacialSimu_calculationHandler.py
 BifacialSimu_src/BifacialSimu/Handler/BifacialSimu_dataHandler.py
 BifacialSimu_src/BifacialSimu/Handler/BifacialSimu_radiationHandler.py
 BifacialSimu_src/BifacialSimu/Handler/BifacialSimu_spectralAlbedoHandler.py
 BifacialSimu_src/BifacialSimu/Handler/BifacialSimu_spectralAlbedoHandler_1_row.py
+BifacialSimu_src/BifacialSimu/Handler/__pycache__/BifacialSimu_calculationHandler.cpython-39.pyc
+BifacialSimu_src/BifacialSimu/Handler/__pycache__/BifacialSimu_dataHandler.cpython-39.pyc
+BifacialSimu_src/BifacialSimu/Handler/__pycache__/BifacialSimu_radiationHandler.cpython-39.pyc
+BifacialSimu_src/BifacialSimu/Handler/__pycache__/BifacialSimu_spectralAlbedoHandler.cpython-39.pyc
+BifacialSimu_src/BifacialSimu/Handler/__pycache__/BifacialSimu_spectralAlbedoHandler_1_row.cpython-39.pyc
 BifacialSimu_src/Data_conversion/10minute_Flughafen.csv
 BifacialSimu_src/Data_conversion/10minute_to_minute_radiation.py
 BifacialSimu_src/Data_conversion/15minute_to_hourly_albedo.py
@@ -84,6 +72,7 @@ BifacialSimu_src/Lib/readme_pics/Module_Parameter_mode.png
 BifacialSimu_src/Lib/readme_pics/Radiance_Readme.png
 BifacialSimu_src/Lib/readme_pics/Sim_Ctrl_enddate.png
[…]
