+++ Paper/Paper_Checklist.txt
@@ -1,34 +1,0 @@
-
-To-Do:
-
-Software wise:
-- Paper.md has to be in default branch (not in a folder)
-- Software must be open source
-- repository must be clonable without registration
-- Be stored in a repository that is browsable online without registration.
-- Have an issue tracker that is readable without registration.
-- Permit individuals to create issues/file tickets against your repository.
-- Does it Work with GitHub Action? (https://github.com/marketplace/actions/open-journals-pdf-generator)
-
-Paper wise:
-- The paper should be between 250-1000 words.
-- Mention (if applicable) a representative set of past or ongoing research projects using the software and recent scholarly publications enabled by it. there are two publications, where bifacialSimu is used, but not named and three Abstracts submitted to the WC-PEC22
-- Citations to entries in paper.bib should be in [rMarkdown](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html) format.
-- Authors are strongly encouraged to include an automated test suite covering the core functionality of their software. (GitHub Acions f.e.?)
-
-Content:
-Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
-A statement of need: Does the paper have a section titled ‘Statement of Need’ that clearly states what problems the software is designed to solve and who the target audience is?
-State of the field: Do the authors describe how this software compares to other commonly-used packages
-
-
-Done:
-
-
-
-
-
-
-All done?
-Submit:
-https://joss.theoj.org/papers/new
