### 2022-07-28 #4443 @rkurchin
@EwaGomez, that's wonderful; I'm really glad to hear that as it's exactly what we strive for here at JOSS! 😊

Can you give a ping here when those last issues are resolved?
