### 2022-08-02 #4443 @EwaGomez
@rkurchin, One of my co-authors asked to be cut out of the contribution since she does not have time to contribute anything to the changes during the review possible. Am I still able to change the authors?
