### 2022-08-04 64c2267
Added Threading to GUI
stopSimulation does not work yet
