### 2022-08-05 ffa5ef9
Changes in Paper with feedback from JOSS reviewers
