+++ paper.md
@@ -21,16 +21,16 @@ bibliography: paper.bib
 
 # Summary
 
-Due to the ongoing anthropogenic climate change and a global effort to switch energy production from fossil sources to renewable energies, bifacial photovoltaic (PV) markets are recently growing  [@Fischer.April2021]. In this context, the consideration of bifacial PV in simulation programs is gaining importance in research and engineering sectors. Bifacial PV can absorb solar radiation arriving on the front side as well as on the rear side of a PV module and use it to generate electricity, which increases the overall efficiency and the energy yield of the PV module on the same area compared to monofacial PV. Because of the significant impact of the rear side irradiance on energy yield in bifacial PV, the inclusion of albedo, which indicates the light reflectivity of a surface, usually the ground, is indispensable. The bifacial simulation program BifacialSimu, developed at the University of Applied Sciences Cologne (Germany) and programmed in python, calculates the energy yield of a bifacial PV system based on two different radiation models – view factor [@Anoma.2017] and raytracing [@Deline.2017]. Moreover, it offers the option to use a constant, time-variable [@Chiodetti.June2016] or geometric spectral albedo [@Ziar.2019] to calculate the light reflectivity of the ground surface. The energy yield can be calculated using two different electrical models [@electrical_characterization].
+Due to the ongoing anthropogenic climate change and a global effort to switch energy production from fossil sources to renewable energies, bifacial photovoltaic (PV) markets are recently growing  [@Fischer.April2021]. In this context, the consideration of bifacial PV in simulation programs is gaining importance in research and engineering sectors. Bifacial PV can absorb solar radiation arriving on the front side as well as on the rear side of a PV module and use it to generate electricity, which increases the overall efficiency and the energy yield of the PV module on the same area compared to monofacial PV. Because of the significant impact of the rear side irradiance on energy yield in bifacial PV, the inclusion of exact albedo simulation, which indicates the light reflectivity of a surface, usually the ground, is indispensable. The bifacial simulation program BifacialSimu, developed at the University of Applied Sciences Cologne (Germany) and programmed in python, calculates the energy yield of a bifacial PV system based on a variety of simulation models, which the user is able to select. Moreover, it offers the option to use a constant, time-variable [@Chiodetti.June2016] or geometric spectral albedo [@Ziar.2019] to calculate the light reflectivity of the ground surface. The energy yield can be calculated using two different electrical models [@electrical_characterization].
 
 # Statement of need
 
[…]
