+++ paper.bib
@@ -157,3 +157,11 @@ year = {2020}
  address = {Satigny},
  publisher = {{PVsyst SA}}
 }
+
+@article{SAM_NREL,
+author = {Freeman, Janine M and DiOrio, Nicholas A and Blair, Nathan J and Neises, Ty W and Wagner, Michael J and Gilman, Paul and Janzou, Steven},
+doi = {10.2172/1440404},
+mendeley-groups = {Sky / irradiance models},
+title = {{System Advisor Model (SAM) General Description (Version 2017.9.5)}},
+url = {https://www.osti.gov/biblio/1440404}
+}
