### 2022-08-05 cd72e9c
Revert "Added Threading to GUI"
This reverts commit 64c226751269ca5be07f0b9f4ef5fb9244249cb2.
