### 2022-08-05 cire-thk/BifacialSimu#10 @EwaGomez
Thank you for all your comments. I implemented every single one of them and updated the paper!

I totally agree that the selling-point of BifacialSimu is not the  GUI, but the different albedo and different electrical models, I tried to underline that!
