### 2022-08-05 cire-thk/BifacialSimu#15 @EwaGomez
@sidihamady - You are totally right, the novelty of BifacialSimu was not the GUI, but the albedo and electrical models. I made those changes to the original paper!

I still like the idea of Threading and I have been working on it for the last few days. I was able to integrate Threading and to make the GUI responsive during simulations, but unfortunately is matplotlib not threadsafe and was not able to communicate with the Thread, where the calculations were running. Matplotlib is a crucial part of BifacialSimu, since, besides the results data, you get 9 interactive plots at the end of the simulation. So I had to delete Threading again to keep Matplotlib running.
I am now reading into multiprocessing instead of Threading.
By any chance, do you have any experience in Tkinter, Multiprocessing and Matplotlib?

Do you still think that this is an issue for the JOSS publication?
