### 2022-08-05 cire-thk/BifacialSimu#15 @sidihamady
Dear @EwaGomez

To implement the threading you need to understand some important points.
First the GUI (so Tk and matplotlib) runs in the main thread. Tk and matplotlib are not thread safe. And they don't need to be thread safe. The organization is as follows:
1/ the main thread with UI
2/ a working thead, with only calculations and without any direct call to the GUI.
The communication between the threads can be done using various methods, e.g. events, mutex, ... (with a timer for example).
The simplest way is to perform all calculations is a working thread and setup a termination function that can create/update plots. If you try to plot directly from the working thread, the program will not work.
If by design you have mixed UI stuff with the simulation routines, it will take some time to redesign the application. One golden rule is to never mix UI and calculations. Different routines, different classes. It can take some time in the beginning but it is really crucial.
You can look at the application I developed to simulate solar cell I-V characteristic: [https://github.com/sidihamady/Photovoltaic-Model](https://github.com/sidihamady/Photovoltaic-Model).

For JOSS, as I mentioned, if you consider GUI as something new in the publication, then it should follow usability rules, and be always responsive and offers features not offered by a command-line program.
If the novelty is only in albedo and electrical models, then this should be highlghted clearly in the paper, and therefore remove the part relating to the novelty which would be linked to the GUI.
