### 2022-08-07 cc58706
improving importing methods
preventing ImportErrors inside some .py files
