### 2022-08-07 54f94f2
Update setup.py
specified python version in program classifiers and updated Development status to alpha version
