### 2022-08-10 97100b8
creating first unittest file (still has some errors)
