+++ BifacialSimu_src/BifacialSimu/Handler/BifacialSimu_calculationHandler.py
@@ -393,7 +393,7 @@ class Electrical_simulation:
         
                 
         # Create dataframe with data
-        p_bi_df = pd.DataFrame({"timestamps":df_report.index, "P_bi ": P_bi_hourly_average, "P_m ": P_m_hourly_average})
+        p_bi_df = pd.DataFrame({"timestamps":df_report.index, "P_bi": P_bi_hourly_average, "P_m": P_m_hourly_average})
         p_bi_df.set_index("timestamps")
         p_bi_df.to_csv(resultsPath + "electrical_simulation" + datetime.datetime.now().strftime("%Y-%m-%d-%H-%M") + ".csv")
         
@@ -2037,7 +2037,7 @@ class Electrical_simulation:
                 
                     
             # Create dataframe with average data
-            p_bi_df = pd.DataFrame({"timestamps":df_report.index, "P_bi ": P_bi_hourly_average})
+            p_bi_df = pd.DataFrame({"timestamps":df_report.index, "P_bi": P_bi_hourly_average})
             p_bi_df.set_index("timestamps")
             p_bi_df.to_csv(resultsPath + "electrical_simulation" + datetime.datetime.now().strftime("%Y-%m-%d-%H-%M") + ".csv")
             
