### 2022-08-11 038c9b9
Update GUI.py
Deleted space for naming in csv "P_bi" for easier search
