### 2022-08-11 3c71ee4
Update test_test.py
solved some previous errors with simulateViewFactors()

new Error: AssertionError at the same Function level
