### 2022-08-11 46a8493
Update BifacialSimu_calculationHandler.py
updated some imports and added a return value that will help with testing
