### 2022-08-11 7b3b826
Update GUI.py
found missing value in SimulationDict
