### 2022-08-12 13181f8
Update test_test.py
Changed test values to match values in default simulation of GUI
made some adjustments and removed unnecessary values
