+++ BifacialSimu_src/BifacialSimu/Handler/test_test.py
@@ -34,72 +34,49 @@ simulationDict={
     'startHour' : (2019, 11, 1, 0),  # Only for hourly simulation, yy, mm, dd, hh
     'endHour' : (2019, 11, 16, 0),  # Only for hourly simulation, yy, mm, dd, hh
     'utcOffset': -7,
-    'tilt' : 10, #tilt of the PV surface [deg]
+    'tilt' : 25, #tilt of the PV surface [deg]
     'singleAxisTracking' : True, # singleAxisTracking or not
     'backTracking' : False, # Solar backtracking is a tracking control program that aims to minimize PV panel-on-panel shading 
     'ElectricalMode_simple': 0, # simple electrical Simulation after PVSyst, use if rear module parameters are missing
     'limitAngle' : 60, # limit Angle for singleAxisTracking
     'hub_height' : 1.3, # Height of the rotation axis of the tracker [m]
     'azimuth' : 180, #azimuth of the PV surface [deg] 90°: East, 135° : South-East, 180°:South
-    'nModsx' : 1, #number of modules in x-axis
+    'nModsx' : 5, #number of modules in x-axis
     'nModsy' : 1, #number of modules in y-axis
     'nRows' : 3, #number of rows
     'sensorsy' : 5, #number of sensors
-    'moduley' : 2 ,#length of modules in y-axis
-    'modulex' : 1, #length of modules in x-axis  
+    'moduley' : 1.98 ,#length of modules in y-axis
+    'modulex' : 0.992, #length of modules in x-axis  
     'fixAlbedo': False, # Option to use the fix albedo
     'hourlyMeasuredAlbedo' : True, # True if measured albedo values in weather file
     'hourlySpectralAlbedo' : True, # Option to calculate a spectral Albedo 
     'variableAlbedo': False, # Option to calculate sun position dependend, variable albedo
-    'albedo' : 0.26, # Measured Albedo average value, if hourly isn't available
+    'albedo' : 0.247, # Measured Albedo average value, if hourly isn't available
     'frontReflect' : 0.03, #front surface reflectivity of PV rows
     'BackReflect' : 0.05, #back surface reflectivity of PV rows
     'longitude' : -105.172, 
     'latitude' : 39.739,
-    'gcr' : 0.35, #ground coverage ratio (module area / land use)
+    'gcr' : 0.45, #ground coverage ratio (module area / land use)
     'module_type' : 'NREL row 2', #Name of Module
     }
 
-# simulationParameter = {
-# 'n_pvrows': simulationDict['nRows'], #number of PV rows
-# 'number_of_segments': simulationDict['sensorsy'], #number of segments for each PVrow
-# 'pvrow_height': simulationDict['clearance_height'], #height of the PV rows, measured at their center [m]
-# 'pvrow_width': simulationDict['moduley'], #width of the PV panel in row, considered 2D plane [m]
-# 'pvmodule_width': simulationDict['modulex'], #length of the PV panel in row, considered 2D plane [m]
-# 'surface_azimuth': simulationDict['azimuth'], #azimuth of the PV surface [deg] 90°= East, 135° = South-East, 180°=South
-# 'surface_tilt': simulationDict['tilt'], #tilt of the PV surface [deg]
-# 'albedo': simulationDict['albedo'], # Measured Albedo average value
-# #'a0': albedo, # Measured Albedo under direct illumination with a solar zenith angle of approx. 60°
-# #'adiff': albedo_diff, # Measured Albedo under 100% diffuse illumination
-# 'C': 0.4, #Solar angle dependency factor    
[…]
