### 2022-08-12 9e2e52f
Update BifacialSimu_radiationHandler.py
commented out this plot because it was causing an AssertionError in the Test
(haven't figured out the reason yet)
