### 2022-08-15 7503f7f
Update BifacialSimu_calculationHandler.py
added a return value which is necessary for testing

and detected a missing variable defined here but not moduleDict (was noted in line 1382)
