### 2022-08-15 b91d12a
Update BifacialSimu_radiationHandler.py
reorganized imports to be read and distinguished easier
