### 2022-08-15 e2faf26
Create test_BifacialSimu_radiationHandler.py
new function tests added

still need to figure out how to evaluate the results with the assertEqual function in unittest.
