### 2022-08-15 5145d3d
Implementation of working threading in GUI
