+++ BifacialSimu_src/BifacialSimu/Handler/BifacialSimu_calculationHandler.py
@@ -302,16 +302,17 @@ class Electrical_simulation:
         # Plot total qinc front and back for every row
       
         
-        f, (ax1) = plt.subplots(1, figsize=(12, 3))
+        f = plt.Figure(figsize=(12, 3))
+        ax1 = f.subplots(1)
         ax1.locator_params(tight=True, nbins=6)
-        plt.plot(P_bi_hourly)
+        ax1.plot(P_bi_hourly)
         ax1.set_title('Bifacial output Power hourly')
         ax1.set_xlabel('Hour')
         ax1.set_ylabel('W')
         f.savefig("P_bi_hourly" + datetime.datetime.now().strftime("%Y-%m-%d-%H-%M") + ".png", dpi = dpi)
-        plt.show()
+        #plt.show()()
         #sns wird entfernt da die Plots in der GUI ansonsten nicht richtig angezeigt werden können außerhalb der Konsole
-        #plt.show(sns)
+        ##plt.show()(sns)
         
         
         ####################################################
@@ -599,15 +600,16 @@ class Electrical_simulation:
         # Plot total qinc front and back for every row
       
         
-        f, (ax1) = plt.subplots(1, figsize=(12, 3))
+        f = plt.Figure(figsize=(12, 3))
+        ax1 = f.subplots(1)
         ax1.locator_params(tight=True, nbins=6)
-        plt.plot(P_bi_hourly)
+        f.plot(P_bi_hourly)
         ax1.set_title('Bifacial output Power hourly')
         ax1.set_xlabel('Hour')
         ax1.set_ylabel('W')
         f.savefig("P_bi_hourly" + datetime.datetime.now().strftime("%Y-%m-%d-%H-%M") + ".png", dpi = dpi)
-        plt.show()
-        #plt.show(sns)
+        #plt.show()()
+        ##plt.show()(sns)
          
         ####################################################
         # Monofacial performance Calculation
@@ -872,7 +874,8 @@ class Electrical_simulation:
                        
                        
                        #The P-V and I-V curve is plotted with the values since the last Rs, Rp adjustment
-                       f, (ax1) = plt.subplots(1, figsize=(6, 6))
+                       f = plt.Figure(figsize=(6, 6))
+                       ax1 = f.subplots(1)
                        ax1.locator_params(tight=True, nbins=6)
                        ax1.plot(Vfplt[-round(V_oc_f0 * 10):], Ifplt[-round(V_oc_f0 * 10):], color = "red")
                        ax1.set_title('P-V I-V Curve front side', fontsize=14)
@@ -882,7 +885,7 @@ class Electrical_simulation:
                        ax2.plot(Vfplt[-round(V_oc_f0 * 10):], Pfplt[-round(V_oc_f0 * 10):], color = "blue")
                        ax2.set_ylabel('Power [W]', fontsize=14, color = 'blue')
                        f.savefig("P-V_I-V_Curve_front" + datetime.datetime.now().strftime("%Y-%m-%d-%H-%M") + ".png", dpi = dpi)
-                       plt.show()
+                       #plt.show()()
  
                        #Rs and Rp have been calculated so the script quits the loop 
                        break
@@ -977,7 +980,7 @@ class Electrical_simulation:
                        ax2.plot(Vrplt[-round(V_oc_r0 * 10):], Prplt[-round(V_oc_r0 * 10):], color = "blue")
                        ax2.set_ylabel('Power [W]', fontsize=14, color = 'blue')
                        f.savefig("P-V_I-V_Curve_back" + datetime.datetime.now().strftime("%Y-%m-%d-%H-%M") + ".png", dpi = dpi)
-                       plt.show()                       
+                       #plt.show()()                       
                        
[…]
