### 2022-08-16 cire-thk/BifacialSimu#15 @EwaGomez
Dear @sidihamady,
Thank you for your input!
I just released a new version of BifacialSimu including threading. The GUI is now responsive during the simulation. But I also implemented your second suggestion and shifted the main USP of the paper from the GUI to the different simulations models, especially the albedo models. In the future, after the JOSS publication, I will implement a timer and a function to update the plots.
