### 2022-08-16 cire-thk/BifacialSimu#7 @EwaGomez
Dear @kanderso-nrel,
Thank you for your suggestion. I just launched a new version of the programm with Tests for the most important functions in the calculationHandler and radiationHandler. I will definitly continue to write and implement tests for the other functions.
You can find the Tests in the folder Tests.
