### 2022-08-16 b21c77f
Implementation of Tests in GUI
