### 2022-08-16 c17893c
Update test_BifacialSimu_simulationController.py
added descriptions of tested functions
