### 2022-08-17 e739536
Revert "Merge branch 'Project_improvements_Aoun'"
This reverts commit b6a93047d6ce9627fb276f6242a6ab1795248745, reversing
changes made to 1cb89981a0a31b21b7ba76f16258300158afabee.
