### 2022-08-18 40c2286
added an additional return value to be used in the Tests
