### 2022-08-18 4066442
Update test_BifacialSimu_radiationHandler.py
