### 2022-08-24 #4443 @sidihamady
@rkurchin , @EwaGomez my reviewing is complete. As a last step before recommending acceptance of the paper, I would suggest to write a short separate document ("Community guidelines") including a diagram or a description of the software architecture (as I suggested in the prereview stage) and a short guideline on how to contribute to the development.

I did not check the "Automated tests" item since there is no really such a test in this software, but the given step-by-step example is sufficient.
