### 2022-09-03 cire-thk/BifacialSimu#19 @kanderso-nrel
[JOSS] egg-info should be removed
The repository currently has a `BifacialSimu.egg-info` directory.  I suggest deleting it (and potentially adding it to the `.gitignore`) since it is an artifact of the build and not actually part of the repository's source code.

Cross-reference: https://github.com/openjournals/joss-reviews/issues/4443
