### 2022-09-03 cire-thk/BifacialSimu#20 @kanderso-nrel
[JOSS] package version in setup.py is out of date
I see that the latest version on [PyPI](https://pypi.org/project/BifacialSimu/) is `1.0.2`, but `setup.py` still lists `1.0.0`:

https://github.com/cire-thk/BifacialSimu/blob/158ca370ff6c34f4381f59bed634dd84259b685a/setup.py#L29

Something to consider: rather than manually updating a hardcoded version string in `setup.py`, what many projects do is to determine the version automatically from the git repository information, in particular git tags.  [setuptools_scm](https://github.com/pypa/setuptools_scm) or [versioneer](https://github.com/python-versioneer/python-versioneer) are two ways to do this.  Using git tags as the basis for versioning is also nice because it makes it easy to go back to the version of the code for older releases.  For example here is the code for pvlib version 0.7.1: https://github.com/pvlib/pvlib-python/tree/v0.7.1, which is easy to find because of the git tag.

Cross-reference: https://github.com/openjournals/joss-reviews/issues/4443
