### 2022-09-04 594dedf
Test still in progress
manual result comparison was added to detect the reason of test failure.
Outcomes are giving some illogical comparison results. (Giving a False for comparison when 2 values are identically equal)
