### 2022-09-04 05707eb
adding dummy variables
since the variable was used in the testing, it is still necessary to add it whenever the function needs to be called
so here I had to add a dummy variable to take this value and avoid the error message.
(might remove this variable later from the test in case it is considered unnecessary, and to avoid confusion by users using this function
)
