+++ BifacialSimu_src/GUI.py
@@ -17,7 +17,7 @@ overview:
     Command to run the simulation
     
 """
-
+# Test
 # Import modules
 #import matplotlib
 #matplotlib.use("TkAgg")
