### 2022-09-07 cire-thk/BifacialSimu#19 @EwaGomez
Thank you!
We solved this issue in the latest push.
