### 2022-09-08 0a9b3e2
Added contribution guidelines and architecture diagram
