### 2022-09-08 #4443 @EwaGomez
Dear @sidihamady,
I added another figure with the general architecture and functions in the ReadMe file in the section "Introduction".
In the section "Contribution", I already mentioned that people should contact me if they want to further develop BifacialSimu. I added a few ideas on how to contribute and I will improve the guide in the future and convert it to a separate contribution guideline. I have oriented myself to the guidelines of bifacial_radiance, which are very helpful!
