### 2022-09-09 cire-thk/BifacialSimu#21 @EwaGomez
Dear @kanderso-nrel,
Unfortunatly I was not able to reproduce your error.
Nevertheless, I changed the two lines, you mentioned and pushed it. In both cases the year is included in the dataframe.
Does that solve your error?
