### 2022-09-09 6223780
Issue #21 ViewFactors error
 [JOSS] ViewFactors error: "dateutil.parser._parser.ParserError: Unknown string format: 11-01 00:00%" #21
