### 2022-09-11 cire-thk/BifacialSimu#21 @kanderso-nrel
> Does that solve your error?

Unfortunately not -- it seems 6223780 only removed the extra `%` in the tracking case, but it is still there in the fixed-tilt case.

But I wonder why you don't get the same errors I do?  This doesn't seem like something where the OS would matter (unfortunately I cannot test on Windows at the moment).  Do you not get errors from this snippet?

```python

In [1]: import pandas as pd

In [2]: pd.__version__
Out[2]: '1.3.4'

In [3]: pd.to_datetime('11-01 00:00%')

  File "/tmp/ipykernel_236209/955505565.py", line 1, in <module>
    pd.to_datetime('11-01 00:00%')
...
ParserError: Unknown string format: 11-01 00:00%

In [4]: pd.to_datetime('11-01 00:00')

  File "/tmp/ipykernel_236209/3743771804.py", line 1, in <module>
    pd.to_datetime('11-01 00:00')
...
OutOfBoundsDatetime: Out of bounds nanosecond timestamp: 1-11-01 00:00:00
```
