### 2022-09-11 cire-thk/BifacialSimu#21 @EwaGomez
@kanderso-nrel
Unfortunately I am still not able to reproduce the error. I even tried a fresh installation of BifacialSimu on OS and it worked.
The only error I get on OS is "name 'rootPath' is not defined", but it works, when you run the code a second time. Currently, around 20 students are working with BifacialSimu and we didn't see this error before. Would it be possible for you to do a fresh new installation?
