### 2022-09-16 cire-thk/BifacialSimu#21 @kanderso-nrel
Hi @EwaGomez, I was finally able to test things on a Windows computer (my previous tests were all on Linux).  Thank you for your patience during my delays.

Indeed I cannot reproduce the error on Windows either.  The difference is in that trailing `%` symbol again -- it gives different behavior on Linux than it does on Windows.  Here is a simple example:

```python
>>> pd.to_datetime(['2019-11-01 00:00:00-07:00']).strftime('%Y-%m-%d %H:%M%')
Index(['2019-11-01 00:00:00-07:00'], dtype='object')  # Windows
Index(['2019-11-01 00:00%'], dtype='object')  # Linux
```
So when I removed that `%` symbol on Linux it got rid of the timezone, leading to the errors in my last comment.  But on Windows the extra `%` symbol ends up just returning the "full" timestamp string for some reason.  I'm not seeing this behavior documented anywhere in the `strftime` format string specification: https://docs.python.org/3/library/datetime.html#strftime-and-strptime-format-codes

I guess the difference is explained by this comment on that page:

> The full set of format codes supported varies across platforms, because Python calls the platform C library’s strftime() function, and platform variations are common. To see the full set of format codes supported on your platform, consult the [strftime(3)](https://manpages.debian.org/strftime(3)) documentation. There are also differences between platforms in handling of unsupported format specifiers.

Between this issue and #18, maybe the easiest way forward is to say in the README that Linux is not supported?  I can finish the JOSS review only using Windows if so.  Let me know what you think.
