### 2022-09-19 95bcc33
Added section, that Linux is not supported at the moment
