### 2022-09-19 cire-thk/BifacialSimu#21 @EwaGomez
Hi @kanderso-nrel,
I agree! I just integrated a part in the README, that Linux is not supported at the moment. But I will leave the issue open, so I can work on it in the next weeks after the JOSS review. I see the importance of BifacialSimu beeing available on Linus, Mac and Windows.
Do you agree with that procedure?
