### 2022-09-19 cire-thk/BifacialSimu#21 @kanderso-nrel
Sounds good 👍
