### 2022-09-19 #4443 @rkurchin
@sidihamady and @kanderso-nrel, from a glance at the discussion here and on issues, it looks like perhaps all of the concerns have been addressed? If not, please make sure @EwaGomez is aware of what other issues you have. If not, it would be great if you can finish up your checklists!

And @sidihamady, to the lack of automated tests point, I think you can still check that box since the full text does read "Are there automated tests or manual steps described so that the functionality of the software can be verified?"
