### 2022-09-23 e22b31a
integrated automated Versioneer (version 1.1.0)
