+++ BifacialSimu_src/__init__.py
@@ -1,1 +1,4 @@
-from BifacialSimu.src.GUI import Window
+from . import GUI, BifacialSimu_main
+
+from . import _version
+__version__ = _version.get_versions()['version']
