### 2022-09-25 #4443 @sidihamady
Hi @rkurchin,
As far as I know, there are no performance claims of the software, this is why I checked off this item.
So there is no action required from @EwaGomez.
Sidi
