### 2022-09-28 #4443 @kanderso-nrel
Thanks @EwaGomez and @rkurchin for your patience.  Here are a few suggestions for the manuscript:
- In the title, "Hollistic" may be a misspelling of "Holistic" (although perhaps this is regional variation?)
- Cuavas 2005 reference: month is "juni" instead of "june"
- Regarding the submissions to WCPEC-8, consider changing "submitted to" to "presented at" now that the conference will soon be over

Otherwise I have completed my checklist and recommend this submission be accepted for publication.
