### 2022-09-29 c023d5e
Last changes in Paper an References before Publication in JOSS
