+++ paper.bib
@@ -82,11 +82,11 @@ year = {2014}
 
 @misc{Cuavas.juni2005,
  author = {Cuavas, Andres},
- date = {juni 2005},
+ date = {june 2005},
  title = {{The early history of bifacial solar cells}},
  address = {Barcelona},
  institution = {{The Autralian National University}},
- year = {juni 2005}
+ year = {june 2005}
 }
 
 @misc{Stein.April2021,
