+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'BifacialSimu: Hollistic Simulation of large-scale Bifacial Photovoltaic Systems'
+title: 'BifacialSimu: Holistic Simulation of large-scale Bifacial Photovoltaic Systems'
 tags:
   - Python
   - Bifacial Photovoltaic
@@ -25,7 +25,7 @@ Due to the ongoing anthropogenic climate change and a global effort to switch en
 
 # Statement of need
 
-Current forecasts expect PV to be the third largest power generation technology by 2050 and will play a significant role in future global electricity generation [@InternationalRenewableEnergyAgengy.2019]. Since the specific yield of a PV-system can be increased with respect to monofacial PV, bifacial PV has become the focus of research within the last decades [@Cuavas.juni2005]. With the rapidly growing market for bifacial PV, there is also a demand for precise simulation programs for this technology to enable planners and investors to receive accurate energy yield forecasts. In recent years, more and more simulation programs and models have been published that allow the simulation of bifacial PV systems [@Stein.April2021]. In contrast to the simulation of monofacial PV, the simulation of bifacial PV is more complex since several factors are added for the energy yield prediction due to the rear side radiation. In addition to the bifaciality and the geometry of the photovoltaic field, these factors include the ground albedo, which indicates the light reflectivity of the ground. There is a need for an open source software tool to simulate bifacial PV, which combines existing radiation, new albedo and new electrical models and also provides a graphical user interface (GUI) to involve every potential user group. BifacialSimu gives the oportunity to choose between various model options for each simulation step and is well documented. Especially advanced albedo models are necessary to simulate in regions, where no measurements are possible.  BifacialSimu has already been applied in two publications [@performance_estimation_bifacial_PV,@albedo_uncertainty_Grommes], three publications regarding the validation were submitted to the World Conference on Photovoltaics Energy Conversion (WCPEC-8) and one publication was submitted to the 17th Conference on Sustainable Development of Energy, Water and Environment Systems (SDEWES) in 2022.
+Current forecasts expect PV to be the third largest power generation technology by 2050 and will play a significant role in future global electricity generation [@InternationalRenewableEnergyAgengy.2019]. Since the specific yield of a PV-system can be increased with respect to monofacial PV, bifacial PV has become the focus of research within the last decades [@Cuavas.juni2005]. With the rapidly growing market for bifacial PV, there is also a demand for precise simulation programs for this technology to enable planners and investors to receive accurate energy yield forecasts. In recent years, more and more simulation programs and models have been published that allow the simulation of bifacial PV systems [@Stein.April2021]. In contrast to the simulation of monofacial PV, the simulation of bifacial PV is more complex since several factors are added for the energy yield prediction due to the rear side radiation. In addition to the bifaciality and the geometry of the photovoltaic field, these factors include the ground albedo, which indicates the light reflectivity of the ground. There is a need for an open source software tool to simulate bifacial PV, which combines existing radiation, new albedo and new electrical models and also provides a graphical user interface (GUI) to involve every potential user group. BifacialSimu gives the oportunity to choose between various model options for each simulation step and is well documented. Especially advanced albedo models are necessary to simulate in regions, where no measurements are possible.  BifacialSimu has already been applied in two publications [@performance_estimation_bifacial_PV,@albedo_uncertainty_Grommes], three publications regarding the validation were presented at the World Conference on Photovoltaics Energy Conversion (WCPEC-8) and one publication was submitted to the 17th Conference on Sustainable Development of Energy, Water and Environment Systems (SDEWES) in 2022.
 
 
 # State of the field
