### 2022-09-29 #4443 @EwaGomez
Thanks @kanderso-nrel - it was indeed a typo!
I made all the suggested changes.

Thank you to both reviewers, @sidihamady and @kanderso-nrel and to @rkurchin - I learned so much in this process and developed BifacialSimu into a better and easier-to-use version. Now it will be difficult to go back to the old school of non-visible scientific review.
