### 2022-09-29 cire-thk/BifacialSimu#20 @EwaGomez
Thank you for your suggestion. Automated packaging is now integrated!
