### 2022-09-29 #4443 @EwaGomez
Thank you for the instructions.
I just published version 1.1.1 under this doi: 10.5281/zenodo.7125653
