+++ paper.bib
@@ -15,6 +15,7 @@
  url = {https://sunpower.github.io/pvfactors/concepts/index.html},
  urldate = {13.09.2021},
  journal = {{44th IEEE Photovoltaic Specialist Conference}}
+ doi = {10.1109/PVSC.2017.8366704} 
 }
 
 @misc{Deline.2017,
