### 2022-09-29 #4443 @EwaGomez
@rkurchin - done. I added the missing doi.
