### 2022-09-29 #4443 @rkurchin
Some cosmetic suggestions:
* lines 12-13: rephrase to "...increases the overall efficiency and energy yield of a PV module of the same area compared to monofacial PV."
* 15: I would move "usually the ground" into parentheses to clarify the flow of this sentence
* 16-17: "programmed in python" reads awkwardly to me, I would remove that and rephrase to "...BifacialSimu, a python package developed at..."
* 23: hyphenate "third-largest"
* 25: don't hyphenate "PV system"
* 26: rephrase to "...has become the focus of research in recent decades"
* 28: I would change "receive" to just "make"
* 32: rephrase to "...several additional factors impact the energy yield due to the rear..."
* 35: hyphenate "open-source"
* 35-36: change to "...which combines existing radiation models with new albedo and electrical models and also provides a graphical..."
* 37-38: rephrase to "BifacialSimu offers various model options for each simulation step..."
* 39: remove comma after "regions"
* 40: something weird is going on with the citation on this line as the formatting is different than the others and you say two publications but there seems to be only one citation there?
* 42: add comma after "(WCPEC-8)"
* 45-49: rephrase to "BifacialSimu is the only open-source bifacial simulation software that combines the two radiation models leading to the most accurate forecast: viewfactor for the front side and raytracing for the rear side. Only one paid option [...] also offers this combination.
* 53: remove "the" before NREL and also "a" before "full"
* 54-57: rephrase to "Neither of the commonly known open-source software packages using the viewfactor model (pvfactors [cite] and bifacialVF [cite]) have energy yield calculations or GUI's, limiting both the target audience and application scope."
* 58: rephrase to "...but the physical models it uses have limited documentation and the software is not free."
* 59-61: rephrase to "To the authors' knowledge, BifacialSimu is the only available software offering all these choices for albedo, radiation, and electrical models."
