### 2022-09-30 92723e1
Added Simulation breaker to running threads
Users could now terminate all long running simulation by pressing the 'Stop Simulation!' button.
