+++ BifacialSimu_src/BifacialSimu/Handler/BifacialSimu_radiationHandler.py
@@ -33,6 +33,9 @@ import numpy
 import dateutil.tz
 import sys
 import math
+from tkinter import messagebox
+from BifacialSimu_src import globals
+
 #import os #to import directories
 #from pvlib.location import Location
 #from tqdm import tqdm
@@ -190,6 +193,14 @@ class RayTrace:
                 
                 for time in range(startHour, endHour+1):
                     
+# =============================================================================
+#                     Check Simulation Break Flag
+# =============================================================================
+                    if globals.thread_break == True:
+                        print("Simulation was Stopped!")
+                        messagebox.showinfo("Simulation Stopped!", "The simulation was successfully terminated!")
+                        break
+                        
                     x = time - startHour
                     
                     singleindex= dtStart + x*datetime.timedelta(hours=1) 
@@ -202,6 +213,14 @@ class RayTrace:
                     
 
                     for j in range(0, simulationDict['nRows']):
+                        
+                        # =============================================================================
+                        #                     Check Simulation Break Flag
+                        # =============================================================================
+                        if globals.thread_break == True:
+                            print("Simulation was Stopped!")
+                            messagebox.showinfo("Simulation Stopped!", "The simulation was successfully terminated!")
+                            break
         
                         key_front = "row_" + str(j) + "_qinc_front"
                         key_back = "row_" + str(j) + "_qinc_back"
@@ -237,7 +256,17 @@ class RayTrace:
             
 
                     for j in range(0, simulationDict['nRows']):
-                    
+                        
+                        # =============================================================================
+                        #                     Check Simulation Break Flag
+                        # =============================================================================
+                        if globals.thread_break == True:
+                            print("Simulation was Stopped!")
+                            messagebox.showinfo("Simulation Stopped!", "The simulation was successfully terminated!")
+                            break
+                        
+                        else:
+                            
                             key_front = "row_" + str(j) + "_qinc_front"
                             key_back = "row_" + str(j) + "_qinc_back"
                             
@@ -319,6 +348,13 @@ class RayTrace:
                 
                 for time in range(startHour, endHour+1):
                     
+                    # =============================================================================
+                    #                     Check Simulation Break Flag
+                    # =============================================================================
+                    if globals.thread_break == True:
+                        messagebox.showinfo("Simulation Stopped!", "The simulation was successfully terminated!")
+                        break
+                    
                     #dataframes to insert results
                     df_rtraceFront = pd.DataFrame()
                     df_rtraceBack = pd.DataFrame()
@@ -356,7 +392,14 @@ class RayTrace:
                     analysis = AnalysisObj(octfile, demo.basename)                   
                    
                     for j in range(0, simulationDict['nRows']):
-        
+                        # =============================================================================
+                        #                     Check Simulation Break Flag
+                        # =============================================================================
+                        if globals.thread_break == True:
+                            print("Simulation was Stopped!")
+                            messagebox.showinfo("Simulation Stopped!", "The simulation was successfully terminated!")
+                            break
+                        
                         key_front = "row_" + str(j) + "_qinc_front"
                         key_back = "row_" + str(j) + "_qinc_back"
                         
