+++ BifacialSimu_src/__init__.py
@@ -1,4 +1,4 @@
-from . import GUI, BifacialSimu_main
-
+from . import GUI
+from . import globals
 from . import _version
 __version__ = _version.get_versions()['version']
