+++ BifacialSimu_src/GUI.py
@@ -58,6 +58,10 @@ import pandas as pd
 #import time
 #import pickle
 import threading #for using multiple threads to make the GUI responsive during simulations
+from BifacialSimu_src import globals
+
+globals.initialize()
+
 
 # aliases for Tkinter functions
 END = tk.END
@@ -152,6 +156,8 @@ ModuleDict = {
     'zeta': 0.06 #Bestrahlungskoeffizient für Leerlaufspannung [-]
     
 }
+
+
 
 
 class Window(tk.Tk):
@@ -373,11 +379,11 @@ class Window(tk.Tk):
         
         # Starting the simulation
         def StartSimulation():
-            
+           
 # =============================================================================
 #             Time Parameter
 # =============================================================================
-
+            
             if (len(Entry_year_start.get())==0 or len(Entry_month_start.get()) == 0 
                 or len(Entry_day_start.get()) == 0 
                 or len(Entry_hour_start.get()) == 0 
@@ -399,12 +405,15 @@ class Window(tk.Tk):
             if int (Entry_month_end.get()) <=0 or int (Entry_month_end.get()) >12:
                 messagebox.showwarning("Simulation Control", "Please insert a End Month between 1 and 12!")
                 exit
+                
             if int(Entry_day_end.get()) <1 or int(Entry_day_end.get()) >31:
                 messagebox.showwarning("Simulation Control", "Please insert a End Day between 1 and 31!")
                 exit
+                    
             if int(Entry_hour_end.get()) <0 or int(Entry_hour_end.get()) >=24:
                 messagebox.showwarning("Simulation Control", "Please insert a End Hour between 0 and 23!")
                 exit
+                    
             if (len(Entry_year_start.get()) != 0
                 and len(Entry_month_start.get()) != 0 and 1 <= int (Entry_month_start.get()) <= 12
                 and len(Entry_day_start.get()) != 0 and 1<= int(Entry_day_start.get()) <=31
@@ -418,15 +427,15 @@ class Window(tk.Tk):
                 SimulationDict["startHour"]=(Startdate.year, Startdate.month, Startdate.day, Startdate.hour)
                 Enddate=datetime.datetime(int(Entry_year_end.get()), int(Entry_month_end.get()), int(Entry_day_end.get()), int(Entry_hour_end.get()))
                 SimulationDict["endHour"]=(Enddate.year, Enddate.month, Enddate.day, Enddate.hour)
-                
+            
             else:
                 messagebox.showwarning("Simulation Control", "Please insert a Start and End Date \n in the format: [yyyy mm dd hh]!")
                 exit
-            
+                
+               
             if len(Entry_utcoffset.get())!=0:
                 SimulationDict["utcOffset"]=float(Entry_utcoffset.get())
            
-
 
 # =============================================================================
 #             Simulation Parameter
@@ -550,7 +559,8 @@ class Window(tk.Tk):
 # =============================================================================
 #             Defining the Path for the Results    
 # =============================================================================
-                
+            
+            
             resultsPath = Controller.DataHandler().setDirectories()
             print('created resultsPath at: ' + resultsPath)     
             
@@ -568,7 +578,7 @@ class Window(tk.Tk):
           
             makePlotAbsIrr(resultsPath)
             makePlotirradiance(resultsPath)
-            #makePlotBifacialRadiance(resultsPath) 
+                #makePlotBifacialRadiance(resultsPath) 
 
           
 # =============================================================================
@@ -626,7 +636,7 @@ class Window(tk.Tk):
                 ax1.xaxis.set_minor_formatter(dates.DateFormatter('%d'))  # day
[…]
