+++ BifacialSimu_src/BifacialSimu/Handler/__init__.py
@@ -3,3 +3,4 @@ from .BifacialSimu_dataHandler import *
 from .BifacialSimu_radiationHandler import *
 from .BifacialSimu_spectralAlbedoHandler import *
 from .BifacialSimu_spectralAlbedoHandler_1_row import *
+

[…]
