+++ paper.bib
@@ -14,8 +14,8 @@
  title = {{View Factor Model and Validation for Bifacial PV and Diffuse Shade on Single-Axis Trackers}},
  url = {https://sunpower.github.io/pvfactors/concepts/index.html},
  urldate = {13.09.2021},
- journal = {{44th IEEE Photovoltaic Specialist Conference}}
- doi = {10.1109/PVSC.2017.8366704} 
+ journal = {{44th IEEE Photovoltaic Specialist Conference}},
+ doi = {10.1109/PVSC.2017.8366704}
 }
 
 @misc{Deline.2017,
