+++ paper.md
@@ -21,7 +21,7 @@ bibliography: paper.bib
 
 # Summary
 
-Due to the ongoing anthropogenic climate change and a global effort to switch energy production from fossil sources to renewable energies, bifacial photovoltaic (PV) markets are recently growing  [@Fischer.April2021]. In this context, the consideration of bifacial PV in simulation programs is gaining importance in research and engineering sectors. Bifacial PV can absorb solar radiation arriving on the front side as well as on the rear side of a PV module and use it to generate electricity, which increases the overall efficiency and energy yield of a PV module of the same area compared to monofacial PV. Because of the significant impact of the rear side irradiance on energy yield in bifacial PV, the inclusion of exact albedo simulation, which indicates the light reflectivity of a surface (usually the ground) is indispensable. The bifacial simulation program BifacialSimu, a python package developed at the University of Applied Sciences Cologne (Germany), calculates the energy yield of a bifacial PV system based on a variety of simulation models, which the user is able to select. Moreover, it offers the option to use a constant, time-variable [@Chiodetti.June2016] or geometric spectral albedo [@Ziar.2019] to calculate the light reflectivity of the ground surface. The energy yield can be calculated using two different electrical models [@electrical_characterization].
+Due to the ongoing anthropogenic climate change and a global effort to switch energy production from fossil sources to renewable energies, bifacial photovoltaic (PV) markets are recently growing  [@Fischer.April2021]. In this context, the consideration of bifacial PV in simulation programs is gaining importance in research and engineering sectors. Bifacial PV can absorb solar radiation arriving on the front side as well as on the rear side of a PV module and use it to generate electricity, which increases the overall efficiency and energy yield of a PV module of the same area compared to monofacial PV. Because of the significant impact of the rear side irradiance on energy yield in bifacial PV, the inclusion of exact albedo simulation, which indicates the light reflectivity of a surface (usually the ground) is indispensable. The bifacial simulation program BifacialSimu, a Python package developed at the University of Applied Sciences Cologne (Germany), calculates the energy yield of a bifacial PV system based on a variety of simulation models, which the user is able to select. Moreover, it offers the option to use a constant, time-variable [@Chiodetti.June2016] or geometric spectral albedo [@Ziar.2019] to calculate the light reflectivity of the ground surface. The energy yield can be calculated using two different electrical models [@electrical_characterization].
 
 # Statement of need
 
@@ -30,7 +30,7 @@ Current forecasts expect PV to be the third-largest power generation technology 
 
 # State of the field
 
-In contrast to other bifacial simulation software, BifacialSimu is the only open-source bifacial simulation software that combines the two radiation models leading to the most accurate forecast: viewfactor for the front side and raytracing for the rear side [@Berrian.October2017]. Only one paid option MoBiDiG hybrid from International Solar Energy Research Center Konstanz also offers this combination [@Berrian.October2017]. SAM by National Renewable Energy Laboratory (NREL) does not include the option to choose between different models for electrial and albedo calculation [@SAM_NREL]. The radiation models in BifacialSimu are based on existing, open-source software. Bifacial_radiance from NREL is used for raytracing [@Deline.2017]. This makes full simulation of a bifacial PV system more complex and time-consuming. Neither of the commonly known open-source software packages using the viewfactor model (pvfactors [@Anoma.2017] and bifacialVF [@Marion.2017]) have energy yield calculations or GUI's, limiting both the target audience and application scope. The commonly used software PVsyst offers an energy yield calculation and a GUI, but the physical models it uses have limited documentation and the software is not free [@PVsystSA]. To the authors' knowledge, BifacialSimu is the only available software offering all these choices for albedo, radiation, and electrical models.
[…]
