### 2022-10-01 #4443 @arfon
@EwaGomez – I found a couple of additional typos in your paper. Could you please merge this PR? https://github.com/arfon/BifacialSimu/pull/1
